upgrades to autocomplete using vectors instead of chat
