+++ app/extension/background.js
@@ -1,4623 +1,11 @@
+import { similarity_all } from './vectorstores.js';
+import { getEditPrompt, getAutoCompletePrompt } from './prompts.js';
+
 function OverleafCursor() {
   // STATE VARIABLES ------------------------------------------------------------
   let msg_chain = []; // contains messages as a dict with keys 'role' and 'content'
   let current_code_snippet = "";
   let api_key = "";
-
-  let average_vectors = {
-    "instructional_creational": [
-      5.796617525629699057e-03,
--1.851442869519814849e-02,
-5.449906317517161022e-03,
-1.238457788422238012e-02,
-7.129017735132947217e-03,
--5.170942527183797037e-03,
--1.437610497814603185e-02,
-1.495667912531644175e-02,
--9.004138741875066659e-03,
-7.454631445762060701e-03,
-2.718488251557573537e-02,
--3.753440540458541450e-03,
--1.996983214630745213e-02,
-5.761987529695033593e-04,
-1.860904416535049538e-02,
-1.214907896355725894e-02,
--4.543463955633342613e-03,
-1.725672208704054399e-03,
-6.430926744360476906e-03,
-2.872622822178527713e-02,
-4.726948680217901884e-03,
-1.324335504323244025e-02,
-1.121622213395312428e-02,
-3.512503986712545156e-02,
--1.272367634810507228e-02,
--2.801628573797643098e-03,
-1.049130356404930423e-02,
-6.494608242064714161e-05,
-2.778905739542096923e-02,
--1.460443960968404965e-03,
-1.332663818029686774e-02,
--3.434890154749155322e-02,
--1.185666428646072780e-02,
--7.385798887116834695e-03,
-9.815228963270783077e-03,
-3.663457625079900165e-02,
-7.634528173366561359e-03,
--1.422424877528101136e-02,
-1.233497498324140872e-02,
--2.269529160112142632e-02,
-2.395674597937613899e-03,
--3.099741650803480159e-02,
-2.186823827214539120e-02,
-1.491388222202658619e-02,
--6.513528060168027878e-03,
--5.468284015660174018e-03,
--7.426518492866307670e-03,
--1.895574297086568410e-02,
-1.117499762913212215e-02,
-2.623199601657688618e-02,
--1.511145411641337012e-02,
--1.810086581390351126e-02,
-9.289690083824098457e-03,
-4.217575641814619129e-03,
--1.564452494785655359e-02,
--6.641266592487227519e-03,
--1.022859493969008327e-02,
-1.775796298607019638e-02,
--2.601276326458901275e-04,
--1.557844026247039507e-02,
-8.685048174811527133e-03,
--1.073264247970655630e-02,
--1.226095517631620255e-03,
-9.939937981835100453e-03,
--2.490740523426211370e-02,
-5.712570438481634830e-03,
--1.399715000297874299e-03,
-1.023547689255792717e-02,
--3.993136750068515539e-03,
-5.374570144340395927e-03,
-3.696296270936727524e-02,
-7.175505138002335678e-03,
-2.363031238201074129e-03,
-1.670656440546736121e-02,
--1.689745357725769281e-02,
--1.567084778507705708e-02,
--2.110887670714873751e-02,
-6.467214412987231861e-03,
-6.047728238627314568e-04,
--1.686844629002735158e-02,
-4.749965574592352087e-03,
-8.718886727001518358e-03,
-1.022013153269654175e-02,
--1.698756285331910407e-02,
--2.506350994808599283e-02,
--7.907167694065719496e-03,
--7.501176645746454597e-03,
--5.707135918783023953e-03,
--6.846516684163361788e-03,
-3.541560639860108495e-03,
--2.800943027250468557e-03,
--1.724777562776580366e-03,
--5.369673774112015963e-03,
-2.039990969933569431e-03,
-7.582265231758355921e-03,
-1.175959423417225445e-02,
--8.937256108038128116e-03,
-2.095185649523045868e-02,
-1.067482004873454527e-03,
-7.633393505238927554e-03,
-1.547256566991563921e-02,
--2.254927053581923246e-02,
--1.459278848105896066e-02,
--2.491052995901554910e-03,
-1.555559545522555709e-02,
-2.070951485948171328e-02,
-5.461339838802814484e-03,
--8.958497748244554151e-03,
--1.420888499851571431e-02,
-1.335785312403459145e-03,
--4.965639067813754082e-02,
-1.175568475155159762e-02,
--1.214827927396982046e-02,
-3.439733937848359069e-02,
--3.475207561859860843e-03,
--1.905171538237482353e-03,
--6.556111450481694365e-03,
-6.528316932963207711e-03,
-1.288634863449260662e-02,
--6.151293526636436235e-03,
-7.642713297514092119e-03,
-1.192349051416386006e-03,
-6.530214968836389917e-03,
--9.021726367063821056e-03,
-2.653775201179086989e-03,
--1.532701669457310276e-02,
--4.361686512129381472e-03,
-5.571368802338838751e-03,
--2.280343815509695490e-02,
-4.670256399549543858e-03,
-7.164363909396342933e-03,
--7.555600046180189003e-03,
-5.735376392840408136e-03,
--2.159228498348966102e-02,
--2.404717153403907945e-02,
--3.006322612054646189e-03,
--1.100781172281131183e-02,
--1.575444481568411079e-03,
--2.213545622071251204e-02,
-1.729515180923044820e-02,
-2.386264976812526445e-02,
-7.559881662018597560e-04,
-3.749854083434911271e-03,
-6.177854456473142256e-03,
-1.463483710540458560e-02,
--6.713384931208566178e-03,
-2.297973600798286584e-03,
-2.019699109951034188e-02,
-1.721985994081478427e-02,
--1.275917384773492778e-02,
--3.255020314827561552e-03,
--3.413928707595914689e-03,
--2.668407331220805714e-02,
-2.525879965978674452e-02,
-2.623235305072739790e-02,
--2.867639869800768862e-02,
-1.163153734814841370e-02,
-2.824230719124898141e-03,
--1.595351808937266538e-02,
-1.524041010998189449e-02,
--9.652123079285956572e-03,
--4.634375325986183833e-03,
--1.434406295884400537e-02,
--1.149794133380055428e-02,
--3.447595582110807386e-03,
-1.287417291314341128e-02,
--1.738429806282510912e-02,
--1.876138477819040348e-02,
--3.023468269966542860e-02,
--1.611129270750098069e-02,
--1.854735251399688284e-02,
--1.705490530875976990e-02,
--1.864780255127698044e-02,
-1.470666842651553383e-02,
-1.802759250276721875e-02,
--9.907823888352141478e-03,
--8.521689895133022219e-03,
-5.114924174267798836e-03,
-6.103426608024165208e-03,
-5.703084834385663618e-03,
-5.343645694665611007e-04,
-2.452055376488715410e-02,
-1.091676646610722015e-02,
--1.334900209621991962e-02,
-6.528030406479956235e-03,
--1.051396827679127420e-02,
-1.397180586063768645e-02,
-2.142519497429020564e-02,
-9.088860406336607412e-03,
-7.965495358803309067e-04,
--2.352004451677203178e-02,
-1.221925042918883299e-02,
-4.551820328924804752e-03,
--4.606010514544322740e-03,
-1.054376597749069264e-02,
--1.329922361765056805e-03,
-2.385141738341189921e-02,
-3.678460442461073572e-03,
--4.945232905447483410e-03,
-1.039587949198903481e-02,
--1.393586429476272384e-02,
--1.157667314691934648e-03,
-8.578030520584433577e-03,
--3.740252481657080439e-03,
--2.472427443135529618e-02,
--1.124669497658032855e-02,
-8.078477272647433200e-04,
--1.169604277238249848e-02,
--2.325532776885665814e-02,
--3.864417371369199914e-03,
-1.937461732304655154e-02,
--9.777826734352856181e-03,
--4.403203478796057987e-03,
-1.029510621447116137e-02,
--6.116808715160004621e-03,
--7.178705319529399102e-03,
-1.445678460877388750e-02,
-2.372098929772619227e-02,
-1.295146818738430669e-02,
-2.977769944700412591e-02,
--1.265686439583077992e-02,
-6.026000317069702197e-03,
-3.468491842504591027e-02,
-1.758764712139964173e-02,
--1.040186439458921051e-02,
-1.538092093978775735e-02,
-2.802045644493773510e-02,
--6.448612429085187787e-03,
-1.940546458354219838e-03,
--1.473289063433185289e-02,
-1.769603228312916984e-02,
--1.129359710030257771e-02,
--1.814531587879173596e-02,
--2.891890692990273237e-02,
--8.746763801900670279e-03,
-4.841169435530900608e-03,
-1.485945822787471116e-02,
--3.383485412341542820e-02,
--1.377483886317349901e-02,
-1.936386675879475661e-03,
--7.477057439973578030e-03,
-3.036079113371670355e-04,
--6.263648712774738669e-03,
--1.413396146381273866e-02,
--1.533920596702955698e-02,
-1.407378227449953591e-02,
--8.664956613210960876e-03,
--1.766654161037877271e-02,
-1.365808543050661770e-03,
-6.283600791357457291e-03,
-1.694241354707628350e-02,
-2.183055127679835866e-02,
--1.250919339727261069e-02,
--1.252481865230947733e-02,
--2.978251185268163820e-02,
--1.036072904244065250e-02,
-2.033802196965552728e-02,
-8.672200108412653555e-03,
-6.992774130776524370e-03,
--2.425668697469518520e-03,
--2.359819386037997901e-02,
--3.650341602042317564e-03,
--1.253444985486567090e-02,
--4.356947931228205562e-03,
--9.402839382528328099e-03,
-1.260561484668869595e-02,
--6.253895326517521988e-03,
--1.938565837917849422e-03,
-1.382548824185505544e-02,
--2.220412745373323560e-02,
--1.063305400311946938e-02,
-3.897202893858775410e-03,
--1.614012100617401193e-04,
--2.942115563200786710e-03,
--2.927287778002209823e-02,
-1.786103069316595862e-02,
--3.668364894110709646e-04,
-2.117743609414901584e-02,
--1.550264834659174050e-02,
-1.362363974021718604e-02,
--1.118172425776720101e-04,
--2.138681791257113132e-03,
-2.525970969581976464e-02,
--4.448930453509092504e-03,
--1.825558947166427883e-03,
-1.574846117291599581e-02,
--9.570746123790741314e-03,
-2.006741430377587815e-02,
-2.218192387226736173e-02,
--3.035549167543649760e-03,
--4.134646014426835228e-03,
--7.504304731264710079e-03,
-1.954220524348784099e-02,
--7.561177469324320381e-03,
-1.307764860393945214e-03,
--4.251396923791617155e-03,
-1.509105544537305867e-02,
-1.748059404781088146e-03,
-6.050235987640917128e-03,
-8.968430437380447290e-03,
-4.266638569533824643e-02,
--1.692886350792832716e-03,
-2.310432039666921074e-03,
-1.110119684599339927e-02,
--1.470325324626173789e-02,
-1.495649204589426587e-02,
--1.364949753042310433e-03,
--1.871450617909431458e-02,
-4.337141360156238253e-03,
-2.144873434444889357e-02,
--1.214606121648103068e-02,
-7.783434493467212070e-03,
-7.989444711711257696e-03,
--2.716578936087898910e-03,
--1.860810421640053322e-02,
--1.971014647278934717e-02,
--5.254452116787433794e-06,
--1.308205677196383476e-02,
--8.060656028101220380e-03,
--1.609163608663948322e-02,
-2.133910732809454064e-02,
--1.585190022597089340e-02,
-1.501140481559559790e-03,
-6.755393452476709908e-03,
-3.599639996537007391e-03,
--1.297420903574675235e-03,
-9.162072590697789565e-03,
-2.010843988973647425e-02,
--5.227086474769748400e-03,
-5.268287379294634299e-04,
--3.020748770795762470e-02,
--1.531897599343210424e-02,
--4.109686799347400318e-03,
-2.177832408342510531e-02,
-2.302088592041400331e-02,
--2.825866050552576850e-02,
--1.766465981490910123e-02,
-2.250399874465074460e-02,
--1.284535592421889340e-02,
--2.465062766859773408e-02,
--2.851521094853524257e-02,
--2.326069672344601574e-02,
--2.018802592065185170e-02,
-4.871028719935566360e-03,
--4.598322772653773600e-03,
--2.537318854592740319e-04,
-1.271044843015261064e-02,
--2.775250613922253251e-03,
-1.345682188402861322e-02,
-1.879977636272087832e-02,
--6.140293236239813095e-03,
-1.099048548494465684e-02,
-1.042440267046913435e-02,
--2.582598768640309789e-04,
-8.190891146659851421e-03,
--1.005660668015479972e-02,
-1.163266415242105757e-02,
--9.137748120701871465e-03,
-1.162268957123160397e-02,
-1.372564080520533025e-02,
--4.686172326910309320e-03,
-4.324932306190021336e-03,
-1.608698774307413398e-02,
-7.848906971048563719e-03,
--1.568682955112308403e-02,
-3.525439702934818429e-03,
--1.125826081261038780e-02,
--1.088547759572975303e-02,
-1.798784224665723672e-02,
-1.472764718864709790e-02,
--2.282942981255473708e-02,
--3.627576640428742567e-03,
--9.521359467180445443e-03,
--1.320442474097944727e-02,
--2.499304831726476689e-02,
-1.561664838372962480e-02,
--5.702905237558297986e-03,
-9.298600576585158706e-03,
-5.782505975366803233e-03,
--2.318792658479651523e-03,
--2.856761532893869852e-03,
-7.604991731932386409e-03,
--4.053207486867904663e-02,
--1.313940071850083768e-02,
-2.906869392609223637e-03,
-2.317868639947846462e-02,
-7.535290718078613108e-03,
--1.381061491556465591e-02,
--2.328195316949859331e-02,
--1.832965267822146346e-02,
-2.418335748370736837e-02,
--6.946946610696614005e-04,
-1.271387308952398598e-02,
-2.520654331310652127e-02,
-2.384966129902750184e-02,
--2.586271623658831405e-03,
-2.729262126376852220e-03,
-3.707870985381304940e-02,
-3.022299036383628915e-02,
-6.023599143372848456e-03,
-1.072731775930151413e-02,
--1.290343486471101685e-03,
--4.825493096723221589e-03,
-5.310113768791780223e-03,
--8.658844052115454923e-03,
-1.317630080156959559e-02,
-2.066508866846561540e-04,
--6.486177764600142837e-03,
-1.863364603650552453e-03,
--3.810103482101112517e-03,
--4.278148870798759336e-03,
--2.584616197273135324e-02,
-3.833749988116323670e-02,
-7.763374211208429404e-04,
-6.731643681996501635e-03,
--2.097129178000613908e-02,
--1.529476239229552413e-02,
-7.959655483136884174e-03,
-2.328894031234085386e-03,
-3.004694385454058717e-02,
-3.606177462643245235e-03,
--1.454659679438918829e-02,
-1.328449659340549295e-02,
--1.377638424455653911e-02,
-7.970367046073078415e-03,
-4.879990788176655908e-02,
-1.976752036716789007e-02,
--2.609046297147870133e-02,
-2.191252362099476228e-02,
--2.817304776981472900e-02,
-7.982799792080186654e-03,
--6.419094155717175454e-03,
--8.865615785180126079e-03,
-1.553628561669029295e-02,
-2.135798454000905483e-02,
--1.330941845662891865e-02,
--1.257099728245520895e-02,
-1.626393728656694293e-03,
-1.392290710937231848e-02,
--1.351727857945661526e-02,
-9.437076002359389565e-03,
--3.808077287976629945e-03,
--2.711962840985506634e-02,
-2.809112305840244517e-03,
--3.530683608260005990e-02,
-5.451034667203202844e-03,
-3.446320185321383352e-03,
-3.684690770751330828e-03,
--5.714306925074197525e-03,
--1.591782353934832056e-02,
-3.102500049863010645e-03,
--1.668012999143684422e-02,
-1.096256765158614213e-02,
--2.227537997532635840e-03,
--1.293648537248373066e-02,
-1.818925312545616223e-03,
-1.627062249463051644e-03,
-3.580137650351389225e-03,
-1.086319654132239521e-02,
-3.980458830483257879e-04,
-2.381921338383108377e-02,
--6.395747841452248666e-05,
--1.411964955041184952e-02,
--3.495535347610712051e-04,
-1.063261677045375039e-02,
--4.387976159341633667e-03,
-1.122590980958193541e-02,
-3.094487498747184803e-02,
--1.520316203241236543e-02,
--4.770191884017549282e-03,
--1.209946066956035706e-03,
--1.420086114403602481e-02,
-2.554979167180135985e-02,
--5.264708632603287350e-03,
--1.222723451210185920e-02,
-4.480484471423551093e-03,
-1.754576637176796794e-02,
--1.447352444520220126e-02,
-1.155453985556960037e-02,
--2.658611501101404428e-02,
--1.830207210514345206e-02,
--7.648764469195157290e-03,
-8.719134202692658251e-05,
--9.253985466784797609e-03,
--2.392416272778064013e-02,
-1.413847412914037791e-03,
-2.874084161594510009e-02,
--3.869942366145551291e-03,
--2.511950454209000046e-03,
--5.760175251634791320e-03,
-2.495110395830124790e-03,
-5.566697020549326788e-03,
-4.787147603929042469e-03,
--1.493360525928437779e-02,
-4.010119551094248717e-03,
--2.349651957629248639e-02,
--2.041575069306418230e-02,
--1.308973871346097440e-02,
-2.502219653542852099e-02,
--1.550656089966651052e-02,
-2.084774865070358058e-02,
--3.334172689937986347e-03,
--2.329609884764067759e-02,
-5.685193311364855617e-03,
-9.969869159976951084e-03,
-4.815942351706325661e-03,
--1.338236943411175162e-02,
-1.325184860615991056e-02,
-2.966446556092705491e-03,
--1.911894799559377177e-03,
--1.368462736718356609e-02,
--1.081534286640817299e-03,
-1.575091123813763400e-02,
-4.471254255622625524e-03,
--5.264019890455528747e-03,
-1.311924971960252116e-02,
-4.450152441859245127e-03,
-2.424953185254707932e-02,
-1.391930615645833348e-02,
--1.286929041962139343e-02,
-1.695476123131811510e-04,
-5.166802112944424499e-03,
-7.603289536200463425e-03,
-1.672262842330383276e-02,
-7.281881989911198616e-04,
-1.647996192332357168e-03,
--8.198171510593965303e-03,
--8.170434973726515074e-03,
--8.895373367704451778e-03,
--1.185641051124548470e-02,
-7.749587940634228116e-03,
--1.081945747428108076e-02,
-6.799012719420715820e-03,
-3.295229605282656675e-03,
--1.462135920883156359e-02,
--1.884451057994738360e-02,
--7.929557083116378635e-03,
--4.369606106774881309e-03,
--5.860209785168990493e-04,
--1.548156049102544698e-03,
--4.918743012240156531e-03,
-1.001555770635604789e-02,
-9.974012962629785761e-03,
--9.891550068277864954e-03,
--2.367893510963767906e-02,
-1.670288061723112973e-03,
--3.462326069711707706e-03,
--4.725664178840816368e-03,
--7.015008942107670183e-03,
-4.463858739472925836e-03,
-1.186602351954206894e-02,
--8.725955695263109754e-03,
-5.642330140108242804e-03,
--1.399469276075251435e-02,
--4.621444107033312494e-03,
--2.862481755437329498e-03,
--2.019495744025334889e-03,
-2.548111279611475939e-03,
-1.085104336962103809e-02,
-1.341262247879058088e-02,
-5.804576590890064836e-03,
--2.845474933565128811e-04,
--2.526359347393736330e-03,
-5.687022814527154402e-04,
--1.071711941185640206e-02,
-2.384089796105399658e-02,
-3.314634956768714065e-03,
-1.611065355828032036e-02,
-9.949295106343925346e-03,
--4.882018256466835737e-04,
--1.056836606003344094e-02,
--8.082255290355533014e-03,
-3.173776676703710109e-03,
--4.597801857162266831e-02,
-1.691830697818659177e-02,
--5.784474243409931486e-03,
-3.837152197957038706e-03,
--6.370661196706350546e-03,
-1.390490976191358606e-02,
--1.174664420541375945e-02,
--1.168178648222237791e-02,
--1.346678823501861258e-02,
--1.065703473286703194e-02,
-4.116579936817288225e-03,
--2.673659931751899352e-02,
-1.588073052698746324e-02,
-2.702373382635414600e-02,
--1.631808921229094310e-03,
--4.396373695635702114e-03,
-1.031808284460567025e-03,
--4.670596463256515563e-03,
-5.170863680541515003e-03,
--2.984021904603650997e-03,
-1.371655571274459431e-02,
--3.756142487691249872e-03,
-4.597167864267249007e-03,
--6.405755602145290616e-04,
-4.200431416393257678e-03,
--7.192240236327052463e-03,
--6.158175115706399430e-03,
-3.663421008968725974e-03,
--8.304970970493740987e-03,
-2.862915903097018773e-03,
--1.554941039648838341e-02,
-4.166639344475697726e-03,
-1.303791684185853263e-02,
--1.479298539925366640e-02,
--1.321118336636573033e-03,
-1.053014227654784987e-02,
--2.957104733795858781e-02,
-5.241138857672922487e-03,
-1.267542790155857732e-02,
-1.703965805791085586e-02,
--2.120623329537920515e-02,
-2.089383517159149051e-03,
-3.830176452174782666e-03,
--4.733009735355153333e-03,
-3.973861818667501385e-03,
-1.574500470887869527e-02,
--4.685095150489359921e-03,
--5.550880625378340483e-03,
--4.782690689899027695e-03,
-4.693126596976071379e-04,
--7.022410270292311733e-03,
--9.154429097543470215e-03,
-9.149173018522560250e-03,
--1.288073395844548975e-02,
--1.120593335945159127e-02,
-1.825010047177784075e-02,
-8.956607757136225353e-03,
-2.172090045642107586e-02,
--3.312020022713113663e-03,
-1.474543026415631218e-03,
--2.638721062103286524e-02,
--3.190050281409639778e-03,
-1.250091320398496499e-02,
-1.259224386885762249e-02,
-7.426320912782102650e-03,
--1.858697098214179277e-03,
--1.311277046916075120e-02,
-5.998533999081701387e-03,
--8.412722416687756430e-03,
-1.111204095650464296e-02,
--4.621763176692184408e-03,
-7.045204634778201580e-03,
-6.511207688527064982e-03,
-8.278169436380267577e-04,
--4.543366999132558877e-03,
--3.392870092648081568e-03,
--1.596753797493875165e-02,
-1.180086709791794463e-03,
--2.896432917623314898e-04,
-9.073354356223717679e-03,
-1.049926365958526661e-02,
-1.695424373028799753e-02,
--1.998397614079294640e-03,
--5.478490318637341261e-03,
--1.806899083312600920e-02,
--1.286341213271953096e-02,
--3.751952567836269563e-03,
--1.411570154014043545e-02,
--1.759783772286027670e-04,
-3.243122488493099895e-03,
--6.083433004096150572e-03,
--1.244546412490308250e-02,
-8.932750963140279055e-03,
--3.095239840331487184e-03,
-8.736097894143313386e-04,
-6.446015217807144100e-03,
-3.497997322119772347e-03,
--3.527987299021333178e-02,
-1.337243474554270471e-02,
-3.290605100919492820e-02,
-7.653349492466076862e-03,
-1.164899180876091157e-02,
--5.104563769418746060e-03,
-1.425187076383735950e-02,
--3.228353764134226506e-02,
--2.234370188671164041e-03,
-7.302129099844023759e-03,
-5.922928926884197046e-03,
-1.578390481881797314e-03,
--8.140425746387337097e-03,
-3.890399538613564887e-03,
-5.285889567312551315e-03,
-3.463036834727972572e-03,
-9.671316441381350854e-03,
-3.488566214218735868e-03,
--5.073072912637144544e-04,
-1.365938619710505052e-03,
--2.319136609439738025e-02,
-6.796660316467750818e-03,
-3.716014721430837935e-03,
-2.004805821925401826e-02,
-1.311212650034576654e-02,
--3.272271019814070037e-02,
-1.678827089490369032e-02,
-7.515548018272965973e-03,
--1.103857957932632451e-02,
--2.237860412569716434e-02,
-3.675343992654234388e-04,
--2.097601408604532307e-03,
--2.094032257955404791e-02,
--1.457972030621021951e-02,
--1.045767102332320099e-02,
--1.854634740739129553e-02,
--3.480865849996917123e-03,
-1.292598578147590230e-02,
-1.181631339713931118e-02,
--1.773267141961696464e-02,
-4.676714295055717056e-03,
--3.662339807488024061e-03,
-2.941017910052323627e-03,
-4.929477910627611328e-03,
--2.357496106997132232e-02,
--2.355300995986908674e-03,
-2.606090041808783791e-04,
--4.838511352136265660e-03,
-2.921751313260756520e-03,
--1.193426515674218466e-02,
-6.770888277969788760e-03,
-1.174248448223806987e-02,
--2.412496072938665609e-02,
--1.390098684787517344e-02,
--1.427402839763089980e-02,
--7.949476753128691814e-03,
--3.805873997043818149e-03,
--6.761940883006900375e-03,
--7.882043370045721531e-03,
-1.213245402323082089e-02,
-3.944646404124796564e-03,
-3.800534486072137778e-03,
-9.063816029811277292e-03,
-3.896055143559351737e-03,
-1.148424434359185398e-02,
-9.636517020408064821e-03,
--1.664632400352274860e-02,
-8.443163882475346110e-04,
--6.233146126032806759e-03,
-1.194518402917310509e-03,
--9.758434596005827882e-03,
--1.379342685686424375e-02,
-1.367676507215946982e-02,
-9.661857620812952518e-03,
-4.916572378715500062e-03,
-1.807452109969744786e-02,
-1.443997639580629809e-03,
-2.330470321940083384e-02,
--1.195659980003256281e-02,
--6.970401918806601316e-03,
--1.340316250862088124e-02,
-1.073738759150728514e-02,
--4.322233400307596077e-03,
-8.037380652967839739e-03,
--2.941576231387443927e-03,
--1.371557583916000952e-02,
--1.934331159573048423e-02,
--3.597830393118783832e-03,
-2.747269669780507545e-02,
-1.026125812495592982e-02,
--5.281669419491663749e-03,
-7.406079338397830898e-03,
--7.009060226937435974e-03,
-2.645432388817425817e-03,
--8.375750982668250105e-03,
-1.407606282737106128e-03,
-7.838466664543375720e-03,
--6.389460596255958254e-03,
--1.106623903615400108e-03,
--1.425330493657384111e-03,
-1.579205456073395908e-02,
--1.173836283851414962e-03,
-1.335191686666803391e-03,
--4.610407289874274284e-03,
-1.440876853448571672e-02,
-1.213912053499370883e-02,
-1.771919702878221826e-03,
-3.934256844513583755e-03,
-2.413212228566408071e-03,
--2.083953709188790526e-04,
--7.239631522679701447e-03,
--1.324547227704897523e-02,
--2.129349188443257004e-02,
-4.682196250723791135e-03,
--5.784643853985471651e-04,
--1.992541558574885061e-02,
-1.333390502259135160e-03,
--3.245280892588198358e-03,
--1.059953144576866214e-02,
--1.017910279842908564e-02,
--1.168333464884199237e-02,
--3.176146143232472458e-04,
--6.567242849268950348e-03,
--1.061842756462283432e-02,
-1.115033323003444821e-02,
-1.339543008361943066e-02,
--2.045161901041865279e-02,
--5.652054920792579790e-02,
--1.928010766860097785e-02,
-8.726327345357276499e-03,
-1.017700194497592713e-02,
--2.989947516471147711e-03,
--5.950995581224560391e-03,
-6.355130113661289432e-04,
-1.720159266842529247e-02,
-1.991965631023049493e-02,
--1.817648181458935211e-02,
-2.461959992069751024e-02,
--2.789860931989096484e-03,
--9.661202668212354617e-04,
-1.181680215340747894e-02,
-1.042987877735868096e-02,
-2.468250499805435712e-03,
--9.323698657681233747e-04,
--1.344433637568727066e-02,
--1.699139422416919704e-02,
-9.194063121140062106e-03,
--1.138285183697007596e-02,
--1.239481536904349888e-02,
--5.156044350587763066e-03,
--3.682181480689905497e-03,
--6.458967851358465850e-03,
-6.534324446693062956e-03,
--1.753680925030494039e-02,
-1.695159597729798545e-02,
-2.861790241149719891e-03,
--3.777636826271191239e-03,
-1.695466792662046057e-02,
--4.801816480176057320e-03,
--4.567357199266553099e-03,
--2.799237379804253578e-03,
--1.024401167524047196e-02,
--3.391794644994661038e-03,
-1.913039042847230922e-04,
--6.093981442973017693e-03,
-1.131188348517753184e-02,
-1.291838556062430143e-02,
--9.677852503955364227e-04,
--3.759365878067910888e-04,
-2.867393323685974056e-03,
-3.277384850662201535e-04,
--1.803303137421608080e-04,
--3.661243116948753422e-03,
--3.358428238425403660e-03,
-6.484273774549365217e-03,
--1.020516794196737499e-02,
-8.799520292086526752e-03,
--8.505989229888655245e-03,
--2.077710015873890283e-03,
--6.153434183215722268e-03,
--2.233383504208177189e-02,
--1.423578610374534072e-02,
-4.059474414680153478e-03,
--1.386003331281244685e-02,
--2.068532621487974982e-02,
-9.031722138752229165e-03,
-1.603302295552566604e-03,
--1.078222209471277858e-02,
--3.650391990959178656e-03,
--2.632351635256782035e-02,
-1.644773001316934824e-03,
--1.079508181428536807e-02,
-5.658421467524021517e-03,
--1.832785492297261953e-03,
-1.377307344810105826e-02,
-3.899088827893137715e-04,
-2.230728042195551271e-03,
-1.199009379524795804e-02,
--9.020237618824467757e-03,
-1.019910561735741863e-02,
-4.584218485979363669e-03,
-1.292483051074668800e-02,
--1.698899422190152178e-02,
-5.538628384238108646e-03,
-1.332225598598597502e-02,
--6.722034799895482311e-03,
--1.819620971800759573e-02,
-1.131170903099700843e-02,
--1.418102880707010696e-02,
-7.990566547960042953e-03,
-9.577941166935489276e-03,
--2.660092071164399212e-03,
-5.719787003272358561e-03,
-1.259143807692453265e-02,
--7.995866003329864941e-03,
--1.924495096318423748e-03,
-4.567825596313923771e-03,
-2.442494441056624124e-02,
--5.574136506766080856e-03,
-8.905629483342636724e-03,
--7.725769595708698034e-03,
--5.180606665089726275e-03,
-1.156509297143202240e-02,
--1.697683532256633043e-03,
-1.487147859879769420e-02,
--6.722068147792015051e-03,
--1.347479757969267597e-02,
-2.222187935213924131e-02,
-5.484314046771032909e-03,
--1.592100519337691280e-03,
-2.828298788517713720e-03,
--1.252480569382896647e-02,
--7.632095261942595417e-03,
-5.696099810302257191e-03,
-1.188563394898665103e-02,
-1.059688066161470522e-02,
-7.511635008268058127e-03,
-2.123554225545376366e-03,
--9.629916504491121271e-03,
-1.887060271110385656e-02,
--3.268449188908562064e-03,
--1.092981463298201526e-02,
-5.771196697605773633e-03,
-1.047232428099960019e-02,
-1.589153751265257736e-02,
-1.445135232643224299e-02,
--1.721123524475842823e-04,
-4.954251297749579126e-03,
--2.866405271925032226e-03,
--1.842092514270916512e-02,
-6.658597765999729398e-03,
-5.606980705488239999e-03,
-1.729583021369762762e-03,
--3.087513154605403629e-03,
--3.375021554529666727e-03,
-9.767023329914081842e-04,
--1.378030627965927055e-02,
-7.970079526421614338e-03,
-1.933384822332300315e-02,
--1.419482048368081450e-02,
-2.559940335922874372e-02,
-5.095498071750625795e-03,
--1.535464568296447327e-02,
-1.346796957077458536e-02,
--2.704776276368647901e-03,
--4.975263882079162051e-03,
--3.698528511449694417e-04,
-1.142253808211535215e-02,
--1.676532293204218219e-02,
-2.384622622048482472e-03,
-1.271323733963072335e-02,
--5.295439762994647373e-03,
--9.488037368282676437e-03,
--1.716490485705435276e-03,
-1.191111255902796949e-02,
--1.532005952321924269e-02,
-2.470784098841250030e-03,
-9.490274253766983054e-03,
--9.172996319648518693e-04,
--3.207148052752017975e-03,
-1.797621091827750163e-03,
--1.096093387459404732e-02,
-2.269345775712281604e-02,
-8.776237501297145388e-03,
-1.308639522176235958e-03,
--5.496350221801549023e-03,
-7.016608316916972399e-03,
--3.679909440688788717e-03,
-3.334742586594074965e-02,
--1.391793821239844034e-02,
-7.889812334906310537e-03,
-1.452049076324328850e-02,
--1.468733337242156354e-03,
--1.294929207069799373e-02,
--1.910927295684814592e-02,
-1.194780238438397646e-02,
--8.596257498720661500e-03,
--1.290255317580886116e-02,
--7.085252250544726675e-03,
--5.452453377074561579e-04,
-7.078204097342677588e-03,
-4.265307867899537260e-03,
-6.613752638804726650e-03,
-1.306331507512368227e-02,
--9.184819867368788221e-03,
-8.237106696469709277e-03,
--8.230708190239965916e-04,
-7.417581923073157321e-03,
--4.536268941592425108e-03,
--2.066067223204299738e-02,
-1.016990005446132342e-02,
-2.737966603599488805e-02,
--9.811091143637895931e-03,
-8.273971511516719363e-03,
-2.308031194843351787e-04,
-1.222904098976869067e-02,
-2.877324307337403297e-02,
-2.723038864787668059e-02,
-5.540819407906382929e-03,
-4.547738453538840986e-03,
--1.480930880643427372e-02,
-8.217316655645845458e-03,
--6.022118299733847033e-03,
-1.051595910539617797e-02,
-1.840539341792464187e-02,
-9.495772555237636664e-03,
--5.803200625814497124e-03,
--4.471408545505255183e-02,
--7.753369596321136106e-03,
-1.917080561397597332e-02,
--1.002544628572650184e-02,
--3.692081433837301856e-03,
-6.569729733746499170e-03,
--9.998067501510377736e-03,
--2.022917566355317970e-02,
--2.364917428349144583e-03,
--1.608614997676340724e-02,
--3.682656324235722336e-03,
-5.284546269103884350e-03,
--9.069818622265302544e-04,
-9.835128797567449857e-03,
-5.298142701758479060e-03,
--8.500066483975388110e-03,
--1.344869594686315438e-03,
-2.246660852688364511e-03,
--8.811316716310102840e-03,
--2.669100314378738265e-02,
-7.110357083729468113e-03,
--1.499923544970442960e-02,
--7.971704122610390186e-03,
-1.569034953863592680e-02,
--1.589355480973608792e-03,
-9.462943362450459972e-03,
--6.699273051344789912e-03,
--5.602400813950225364e-03,
-4.237990448018535972e-03,
-6.003179785329848875e-03,
--7.693252956960350622e-03,
-4.045010427944362164e-03,
-7.336955430946545849e-03,
-7.033339687041006982e-03,
-1.252099855628330261e-02,
--1.254344221961218794e-02,
--3.954024583799764167e-03,
-6.711225549224763675e-04,
-1.931499386409996077e-02,
-6.943811944802291344e-03,
-1.088730854098684999e-03,
--1.915427230269415443e-02,
-3.799385602906113664e-03,
--1.751150429045083071e-02,
-5.824532982660457665e-03,
--9.503573726397008051e-03,
-1.157231287797912990e-02,
-3.043556460761465219e-03,
--2.839520739507861522e-03,
--6.137152740848250687e-04,
--8.193383307661861528e-03,
-2.431746654910966685e-02,
-9.322184824850410917e-03,
--5.398151959525421262e-03,
--4.677552159409970219e-03,
--6.962361792102455919e-03,
-3.413277817890047940e-03,
-7.005924521945416754e-03,
--4.924211346951779325e-03,
-1.232301911804825102e-02,
--1.174405086785554851e-02,
-7.869364225916797642e-03,
-8.855238452088087103e-03,
--1.672563597094267607e-02,
--3.235706628765910951e-03,
--6.946145724214147776e-03,
--1.429918138310313190e-02,
-8.594994549639523029e-03,
-4.065535192785319288e-03,
--7.946485280990600239e-03,
--1.665230806192994489e-02,
-1.032454577507451246e-02,
-2.936761268210830167e-02,
-8.578775660134851586e-03,
-1.068553701043129011e-03,
-4.306666526099433775e-03,
-1.026874519084230997e-03,
--7.209016813430934971e-03,
-5.287298794428352050e-04,
--1.714009028510190408e-02,
--1.382651618769159541e-02,
-4.374966947943903159e-03,
-2.364538768306374480e-02,
--4.843800142407417124e-03,
-1.326157932635396762e-03,
-6.725195022590923707e-03,
--2.539632812258787314e-02,
--1.374768130772281369e-02,
--1.121016717515885795e-02,
-8.881343313260004618e-03,
--6.608059034624602993e-03,
--5.241428248700686146e-03,
--5.854699673363938494e-04,
-1.908551121596247045e-03,
--1.319631490623578499e-02,
--1.065861345850862597e-02,
--1.861637757392600040e-02,
-9.905288421578006819e-03,
--1.420955979847349313e-03,
--1.752225164091214624e-03,
--3.692906320793554187e-04,
-9.196887048892677177e-03,
-5.643012572545558041e-03,
-3.882337361574172865e-04,
-4.876366211101412426e-03,
--5.388246639631688942e-03,
--2.448241795354988426e-02,
--7.727080488984938364e-03,
--5.246105016703949894e-03,
--1.356318454490974476e-02,
-1.398439299373421905e-03,
-2.554410239099524915e-03,
-1.020986695948522538e-03,
-5.633777251932770187e-03,
--2.076780458446592179e-03,
--3.682125423802062956e-03,
-1.047135125045315328e-02,
-1.891816337592899756e-03,
-1.562294759787619201e-03,
--2.570996928261592897e-03,
--2.891882834956049746e-03,
--2.948892320273443883e-03,
-1.845535387765267002e-02,
-9.862118237651883862e-03,
--7.928176014684141376e-03,
--3.600978656322695315e-03,
-1.140261492691934178e-02,
-2.849798585521057161e-03,
--9.376564435660839428e-03,
-1.376633809995837440e-02,
-3.181094009050866589e-03,
--5.872504145736456502e-03,
--2.128788246773183433e-03,
-3.358481131726875901e-03,
--2.238168878830037838e-03,
-1.649128051358274977e-02,
-2.551097777177346784e-03,
-1.191470929916249501e-03,
-2.410954999504610896e-03,
-5.980972063844092537e-03,
--5.809436400886625398e-03,
--5.688105037552304739e-03,
-5.539429068448953331e-03,
-1.828727216270635905e-02,
--4.084493544723954983e-03,
-9.201238554669543407e-03,
--2.878007281105965463e-03,
--1.390208129305392504e-02,
--1.011398109112633323e-02,
-1.446621495415456593e-02,
-8.059684685431420248e-02,
--8.537710318341851581e-03,
--1.142523922026157310e-02,
-1.508427975059021230e-02,
-1.353390741860494018e-02,
-4.140236793318763375e-03,
--5.841054697521030556e-03,
--6.561168044572696263e-03,
--3.452512400690466079e-03,
-9.881640621460973523e-03,
-3.602805919945239934e-03,
-2.262484328821301287e-03,
-1.151000848039984703e-02,
--1.655986066907644218e-04,
-4.366598225897177880e-03,
-5.267424951307475220e-03,
--3.477242484223097563e-03,
-3.851028275676071644e-03,
--3.760833946580533025e-03,
-1.676638546632602724e-03,
-7.231073123693932021e-03,
--1.804119240841828330e-03,
--1.970963779021985690e-02,
--2.079644822515547330e-04,
-4.368397803045809095e-03,
--2.491111033123161178e-02,
-6.473757477942853819e-03,
-7.965298008639366129e-03,
-1.145533910021185806e-02,
--3.096577222459018317e-03,
-1.227273053227690910e-02,
--3.044970717746764508e-03,
-7.878538279328495264e-03,
-4.870644502807408918e-03,
-2.055602055042982015e-03,
-1.495109966490417754e-02,
--8.048338297521695148e-03,
-1.244041780591942413e-03,
--1.533152235206216608e-02,
--3.382333047920837825e-03,
--1.041922785225324333e-02,
-1.769462852389551769e-02,
--8.338434062898159374e-03,
--1.581956270383670995e-02,
-7.043538801372051412e-03,
--8.895977213978768089e-03,
--5.869554675882682204e-03,
--7.234635995700955391e-04,
-7.435561699094250625e-03,
-5.091299329069443221e-03,
--6.951868592295795679e-03,
-8.500982163241133311e-04,
-3.815894652507268066e-03,
-7.298578487825580288e-03,
--1.242994292988441885e-02,
--5.937619518954306494e-03,
--6.377238966524601156e-03,
-7.454221282387152531e-04,
-1.947585944435559213e-02,
-1.424452756764367153e-02,
--7.986368390265851519e-03,
-5.207597919797990818e-03,
--7.213642948772758419e-03,
--3.851610637502744707e-03,
-9.124763624276966203e-03,
-1.335725525859743400e-03,
--7.750403960380935695e-03,
--7.595547265373170723e-03,
--1.435946885612793180e-03,
-2.823012112639844244e-03,
-4.776678040798287947e-04,
--7.198913335741963238e-03,
--7.997743683517910540e-03,
-1.895534060749923949e-02,
--2.834282146432087843e-03,
--5.028680676514340857e-03,
-2.522868890082463622e-03,
-4.457976645971939780e-03,
--9.078716899966820381e-03,
--3.169226444588275904e-02,
--5.475508674862794795e-04,
-9.633517352631315936e-03,
-5.990094761364162315e-03,
-2.134365923120640218e-02,
--1.871616691350936820e-02,
-1.334537792718037920e-03,
--3.874925198033452034e-03,
--9.487883225665428091e-03,
-1.067272495711222256e-02,
-1.072246901458129287e-03,
--6.339679706434253950e-03,
-3.814956330461427743e-03,
--7.996986279613338758e-03,
-5.414858198491856456e-03,
-1.010927454335614969e-02,
-1.166037074017367515e-03,
--6.679935916326940233e-03,
-4.971846922126132648e-03,
--8.854243726818821630e-03,
--4.150034312624483954e-03,
--5.901969072874635111e-03,
--1.130226142413448542e-02,
--5.637266789563000202e-03,
-6.415869179181754763e-03,
-9.444311865081544966e-03,
--6.838096250976377806e-03,
--1.315397438593208720e-02,
--1.459800790835288388e-02,
-1.759244187269359692e-05,
-1.525529617210850085e-02,
-8.307637984398753145e-03,
--5.134074547095223914e-03,
-1.193972408293575427e-02,
--2.630617981194518368e-02,
-1.276167543328483520e-02,
--3.388711658772081137e-03,
--5.349835441302274938e-03,
--6.470361797255463898e-03,
-4.357590484141837459e-03,
--1.273502206313423778e-02,
--2.826674110838212246e-03,
--8.714316417172084800e-03,
-2.630607696482911793e-03,
--9.599541281204437837e-03,
-1.445051566697657178e-02,
--5.656173499301075762e-03,
-2.369379469892009954e-02,
-1.217717584113415727e-02,
-5.275144038751023064e-03,
--4.358094185590744019e-03,
-8.978337340522557497e-03,
-1.727434576023369889e-02,
-9.021741134347394109e-04,
--1.612637584912590683e-03,
--1.090488003683276474e-02,
-1.376436080900020863e-02,
-5.258520078496075993e-03,
--4.192980450170580298e-03,
-8.965932240244001584e-04,
-1.183785482862731442e-02,
-3.281921660527586763e-03,
--1.411698390620586083e-03,
-7.052823289996013045e-03,
--6.072147324448451573e-03,
-7.090752711519598961e-03,
--7.223101812996901572e-03,
-1.175828048726543749e-02,
--1.046180059202015365e-02,
--7.382516475627198642e-03,
--4.749189503490925182e-03,
--6.000529391167219381e-03,
-1.792717189964605426e-02,
--7.674558913640794691e-03,
-4.695618187543004925e-03,
-6.702211569063365893e-04,
--3.644237434491515073e-03,
--1.033689838950522209e-02,
--2.016767857130617042e-02,
--1.179736491758376360e-02,
--1.041610498796217109e-02,
-2.633757647708989767e-03,
--4.382104982505552124e-03,
--1.384289537090808118e-02,
--1.861164376023225625e-02,
-5.025232792831957340e-03,
-5.001214588992297996e-03,
--1.136256502941250870e-02,
-6.901564920553937554e-03,
-2.099117170473618899e-03,
-5.540636921068653464e-03,
-1.707935317535884748e-02,
--2.052379227097844878e-03,
--9.343086869921534560e-03,
-1.236900191288441489e-02,
--1.719059236347675367e-03,
--3.772138687781989748e-03,
-1.303341612219810399e-03,
-9.144622559506387222e-03,
--3.212343479390256107e-03,
-7.212446560151875366e-03,
-3.985009339521639396e-03,
--1.868237275193678285e-02,
-6.921302003320306973e-04,
--1.065346333198249375e-02,
--1.295314126182347571e-02,
-3.457854187581688252e-03,
-1.159115447662770713e-02,
--1.519478065310977434e-02,
--1.538611110299825755e-03,
--5.346782533160876136e-03,
-1.374990237382007693e-02,
--1.448020654643187265e-02,
--3.120754391420632774e-03,
--2.457056682032998543e-04,
--2.164675792737398227e-02,
-2.833869245660025453e-03,
--5.875868024304509380e-04,
--1.551316575787495787e-02,
--1.131678395904600664e-03,
-2.284671657253056938e-03,
-1.639773733331821928e-02,
--1.894340908620506525e-03,
-1.388690870953723788e-03,
-1.163204107433557510e-02,
-6.193396280286833991e-03,
-3.255684069154085678e-03,
-2.186491236207075556e-02,
-1.580242475029081106e-03,
--1.235758251859806443e-02,
-3.578497527632862243e-03,
-2.897067816229537216e-03,
--4.236914595821872354e-03,
-5.544082919368520206e-03,
-3.063535269757266914e-03,
--1.097084127832204147e-03,
-2.291446842136792700e-02,
-5.796344205737114126e-03,
--1.325727737275883642e-03,
--1.354888499481603573e-02,
--3.889002959476783958e-03,
-3.081767936237156391e-03,
--5.888171767583117062e-03,
--7.362659944919869477e-03,
-3.236914318495110084e-03,
--1.402452607871964609e-02,
-5.584271979751065026e-03,
-1.562166747171431745e-02,
--2.234174911573063443e-02,
--4.809402598766610026e-03,
--7.269878149963915001e-03,
--2.147712768055498513e-03,
-2.390060701873153622e-03,
--1.111330356798134657e-03,
--1.304810343426652218e-02,
-2.499690209515392607e-03,
-8.148330578114837755e-03,
--5.530156529857777059e-03,
--4.921446816297248006e-03,
-9.420621837489307621e-03,
--6.255360952491173347e-03,
-4.948420333676040389e-04,
-4.168705904157831918e-03,
--5.771884554997086178e-03,
-4.772766272071749145e-03,
--1.065845549746882205e-02,
-1.398081353836460108e-02,
-1.992765099857933889e-02,
--1.356925985019188409e-02,
--2.534358645789325237e-05,
-1.648864349117502501e-02,
-5.647441232576965939e-03,
-4.766928893513977354e-03,
--1.364394499687477877e-02,
--1.632234873250126925e-03,
--8.606333739589900472e-03,
-2.370678058359771898e-02,
-4.675366671290248459e-03,
--5.577206306043080986e-03,
-5.809848214266821569e-03,
--2.272638151771388861e-03,
--1.943359791766852140e-03,
-5.694103852147236293e-03,
-5.746777227614074945e-03,
--3.585527636460028697e-04,
--1.027415166608989204e-02,
-3.871417662594467618e-04,
--9.018874500179662954e-03,
-7.645880884956568826e-03,
--9.197440145362634559e-03,
--3.840492630843072913e-04,
--1.358944488456472789e-02,
--1.289579930016770959e-03,
--7.774274505209177732e-03,
--1.577770828007487441e-02,
--1.376760528219165246e-02,
-1.473088267375715146e-02,
-5.901797040132805358e-03,
-5.701260070782154972e-03,
-1.263819966698065400e-02,
-8.232183440122753035e-03,
--4.812938420218415732e-03,
--4.795906814979389483e-03,
--1.107448376133106718e-02,
--1.738827064400538802e-02,
-1.283465602609794656e-02,
--1.125362667589797569e-02,
--2.944272587774321350e-03,
-1.140368605920230073e-02,
-4.072258013547980053e-03,
-2.494510167161934164e-03,
--8.362510823644697666e-03,
-3.725501068402081728e-03,
-3.692370775388553652e-03,
-1.051390838110819424e-02,
-1.900155378389172325e-02,
--1.657554370467551053e-03,
--4.353959136642516006e-03,
--6.351615261519327586e-03,
--4.857458872720599174e-03,
--1.516062879818491594e-03,
-2.958839439088478793e-04,
-1.070468481630086968e-02,
--1.099153441318776424e-02,
-1.296066998329479213e-02,
--5.560721125220879912e-03,
-5.706140783149749214e-03,
-6.017951236572116613e-03,
-4.760722484206780791e-03,
--2.010734673240221950e-03,
--1.281055082799866830e-02,
-7.210352126276120370e-03,
--1.516053947852924520e-04,
--9.976351843215524803e-03,
-8.946971761179155908e-03,
-3.567236554226838139e-03,
-7.482951204292476004e-03,
-6.087191373808309436e-03,
--9.978944315662374286e-04,
--1.368339970140368747e-02,
--1.484721433371305466e-04,
-1.914335584588116104e-02,
-1.830344863847131048e-02,
-1.050140458755777217e-02,
--1.216843012953177164e-02,
-1.693036765755095985e-02,
--1.391814692760817707e-02,
--9.117501869332045664e-03,
--9.320581314386799238e-03,
-1.130275099567370524e-02,
--2.963175810873508280e-03,
--7.352670020190998565e-03,
-7.938120095059276321e-03,
-1.443326355365570635e-02,
--1.206941418349742959e-02,
--2.611389709636569110e-03,
--6.179960386361926794e-03,
-8.615528982772957750e-04,
-3.042007997282780880e-03,
-2.630027959094150079e-03,
--1.655967819388024648e-02,
-7.038334134267643268e-03,
-1.605859301635064254e-02,
-1.682738254894502322e-02,
--8.512851985869928978e-04,
-6.472019775537773620e-03,
-2.081528827548027177e-02,
-1.930629986745770882e-02,
--7.835679920390248082e-04,
-4.297574749216437340e-03,
--1.031020794180221786e-02,
-5.501285199716221715e-03,
--1.075396991800516878e-02,
--1.420643374440260301e-02,
--2.466997120063751762e-03,
--1.552146612812066451e-02,
-2.797507075592875567e-03,
-2.990604755468666623e-02,
--1.175725396024063230e-02,
--6.885011239501181816e-03,
--6.058339774608611887e-03,
-2.880668832949595439e-03,
-4.201156715862452810e-03,
--1.146951464124867925e-02,
--4.674054111819714481e-03,
--7.005285646300762892e-03,
-7.956088101491332748e-03,
-5.163993209134787148e-03,
--8.239683663123286567e-03,
--4.099944416157086553e-03,
-4.923624242655933250e-03,
--1.000556879444047852e-03,
--1.089903081301599706e-02,
-1.281619688961654988e-03,
--4.245337353495415300e-03,
--3.106588782975450071e-03,
--7.419646065682172602e-03,
--7.329797342390520325e-03,
-4.115391706000082542e-03,
--2.744856741628609598e-03,
--1.848463969945441862e-02,
-1.275382364401593874e-02,
--2.065717065124772663e-03,
-1.528475178347434883e-02,
--1.647799183847382595e-02,
-3.538132279645651856e-02,
-1.441990606836043333e-02,
--6.495008623460307717e-03,
-1.446284838020801475e-02,
-2.753310423577204260e-03,
-    ],
-    "unfinished_latex": [
-      -3.973259581835008811e-03,
-8.744894462870434274e-03,
--8.055317634716629288e-03,
--7.455401920014992530e-03,
--1.751273257832508534e-02,
-1.859033247019397164e-02,
--2.095822060364298592e-02,
-2.669851644895970821e-02,
--1.979418627452105353e-02,
--3.342758317012339917e-03,
-1.506193616660311861e-03,
-1.006398356985300710e-02,
--1.015973714675055817e-02,
-1.706497468985617230e-02,
-2.065443931496702135e-02,
-3.853589061181992431e-02,
--6.140403423341922289e-03,
-4.162678774446248835e-03,
-4.828265311516588445e-04,
-3.613275947063811916e-02,
-1.099346728879027103e-02,
--3.367625118698924997e-04,
-2.904278510250151227e-02,
-4.450384952360763972e-03,
--6.147134120692498808e-03,
--1.189552596188150381e-02,
--2.196746382978744666e-02,
-1.142448401078581879e-02,
-4.922766225645318350e-02,
--1.422957038885215228e-02,
-2.952824942767619948e-02,
--3.609892074018716812e-02,
-2.216266572213498875e-02,
--3.879444045014679432e-02,
--7.506611669668927789e-03,
-3.744107810198329250e-02,
-1.787485381428268599e-02,
-5.358831491321325519e-04,
--2.952785363886505435e-02,
--1.241250551538541913e-02,
-7.306755555327981363e-03,
-1.529467361979186492e-03,
--4.716435603040735995e-03,
-1.277329750591888954e-02,
-2.542682273779064278e-02,
-1.427434660727158224e-02,
--5.086504568462259929e-02,
--1.143796327523887227e-02,
-2.853804218466393650e-02,
-7.917443204496521850e-03,
--2.364957803802099010e-02,
--6.276517850346862923e-03,
-1.087480059941299190e-02,
-1.766669069183990359e-02,
--1.478109410381875898e-02,
--1.411852211458608551e-03,
-2.858506352640688593e-03,
-3.967653827276081124e-02,
-2.291722809895873139e-02,
-2.009000665930216287e-03,
-9.199097063537919547e-03,
--6.308554112911224712e-03,
--9.577488897775765075e-04,
--2.451680449303239497e-03,
--2.544487451668828865e-02,
--6.693482377158943564e-03,
--1.901679844886530052e-02,
-2.065804268931970000e-02,
-5.487208795966580337e-03,
--6.419890472898259759e-03,
-9.942441206658259728e-03,
-3.646264501148834686e-02,
--2.859469313407316665e-02,
-1.080586735624819959e-02,
-7.148144890379626304e-03,
--1.689901633653789759e-03,
--4.263541516847908774e-02,
-3.498093265807256276e-03,
--4.083726386306807052e-03,
-2.253368667443282972e-02,
-2.504280641733203083e-02,
-2.352550349314697087e-02,
--7.661223749164491718e-03,
--1.995733280273270793e-02,
--2.141489349887706201e-02,
--2.271508588455617428e-02,
--3.110957400058396091e-02,
--7.970536476932466247e-04,
--2.135029069031588872e-02,
--1.615363174350932171e-02,
--7.376620118157007147e-03,
-1.136203194691915927e-02,
--2.649353645974770327e-02,
-2.765877211495535587e-02,
-2.074462210293859157e-03,
-2.204873053124174406e-02,
--1.883032270707190106e-02,
--1.732160617830231716e-02,
-2.236614992725662748e-02,
-9.370663436129688783e-04,
-2.762139044934883853e-02,
--1.552296521549578756e-02,
-1.774471122771501680e-02,
--1.831492233322933386e-02,
-1.848664225253742258e-02,
-1.133294749888591468e-02,
--1.446236744523048470e-02,
--3.690123688429593901e-02,
--3.819424484390765429e-02,
-1.003097087959758937e-02,
-2.435783122782595594e-02,
--2.381120717618614296e-02,
--2.809699791250750561e-02,
-2.886609745910391142e-02,
-7.066190676414407963e-03,
--1.016267060185782611e-02,
-2.805423315148800750e-02,
--1.219403161230729962e-02,
-4.133120295591652393e-03,
--1.520158144994638791e-02,
--1.997057115659117699e-04,
--2.110574464313686067e-03,
-2.240938654867932062e-02,
--5.625425074249505719e-02,
-5.199041659943759094e-03,
--1.721694635052699665e-02,
--7.853147701825945876e-03,
--7.290923979599028999e-03,
--2.261036707204766572e-02,
--1.738484574307221781e-02,
-2.255194803583435853e-02,
-8.489550091326236725e-04,
--5.013549267459893599e-03,
--6.900661801046226521e-03,
--2.434362768835853646e-02,
--2.084533177549019282e-03,
--1.376168632123153657e-02,
-1.211340274894610013e-02,
--3.331455499865114411e-02,
-2.792396319564431975e-02,
-2.727972451393725212e-02,
--1.137276639929041180e-03,
-2.107373339822515845e-02,
-2.057588454335928033e-02,
-1.830648991162888567e-02,
--2.273723261896521027e-03,
--1.929803152161184768e-02,
-1.351142524072201923e-02,
-4.411324090324342077e-03,
--1.704691734630614519e-03,
--6.175023515243083412e-03,
-1.413247554155532340e-02,
--3.042460333090275457e-02,
--4.074956616386770855e-03,
--1.725879134610295434e-02,
--2.190176274161785910e-02,
-1.397514635464176466e-02,
-1.157898288602154935e-02,
-1.116817494003044010e-02,
-7.384492515120655645e-03,
-1.418961353047052364e-02,
--2.963992182776564732e-02,
--2.093353134114295244e-02,
--2.037936307824566098e-02,
--1.645624818047508597e-02,
-3.182987581240013381e-02,
--1.223997086053714087e-02,
-2.495519933290779591e-03,
--5.320763571653515378e-02,
-4.535343381576240063e-03,
--2.041436027502641082e-02,
--1.944463358377106557e-02,
--2.277987955603748629e-02,
-1.928524321992881527e-02,
--1.183865525817964261e-02,
-4.891169200345757244e-03,
--1.773684252984821658e-02,
--1.869004432810470601e-02,
--6.452457618433981960e-03,
-1.498907190398313062e-03,
-3.343018507584929327e-02,
-5.797518910840153417e-02,
--1.128418180560402176e-02,
-1.457973863289225777e-02,
--3.033333667553961364e-03,
--4.012424062239006162e-03,
--1.148171606473624733e-04,
--9.845346826477908292e-03,
-5.389356467640027057e-03,
--1.688954794080927899e-02,
--3.054580911993980477e-02,
-2.922235441801603745e-02,
-1.872595368186012049e-02,
--4.471853582072071558e-03,
-5.528640432748943914e-03,
--1.966052072239108522e-02,
-2.040305837581399967e-02,
--8.859847276471554539e-03,
--2.194737640093080699e-02,
-1.203827609715517680e-02,
--2.528259065002202852e-05,
-3.109157728031277518e-02,
-1.506878216750919854e-02,
--4.009108245372772217e-02,
--1.503079550166148623e-02,
--3.743553836829960346e-02,
--3.348496071994304379e-02,
--1.156956416089087794e-02,
-9.587417333386838089e-03,
--3.091484878823393817e-02,
--9.297358104959130287e-03,
--6.026116339489817966e-03,
-2.803363663260825117e-02,
-1.710467734956182609e-02,
--2.473799009458161821e-02,
-3.801211167592555371e-03,
--3.406440973049029534e-03,
-2.679838360054418583e-02,
-2.099937659222632716e-02,
-2.547775104176252972e-04,
--2.136118641356006195e-02,
-4.712846358306706229e-02,
-5.675952713936567584e-02,
-1.435134616913274000e-02,
--9.094646415906026266e-03,
-2.013558214530348708e-02,
-3.569096239516511559e-02,
--4.260482303798198561e-02,
-1.622531267057638477e-02,
--1.064220830448903143e-02,
-4.038759353570640226e-02,
-2.005523220868781278e-02,
-1.869875285774469419e-03,
--1.200556157709797844e-02,
--8.927008265163748763e-03,
-1.102921260753646562e-02,
--1.141999643296003411e-02,
--6.043970497557893260e-05,
--1.000145886064274237e-02,
--2.781596259446814795e-02,
--2.155645787715911726e-02,
-5.072970921173691576e-03,
-1.027262099378276582e-02,
-1.866447147331200609e-02,
--3.279702056897803968e-03,
-1.524280269513838063e-02,
--2.853577626228798354e-03,
--1.595507248712237786e-02,
--2.190637937746942217e-03,
-1.666411121841520140e-02,
-1.899441269924864034e-02,
-4.048163760453462462e-02,
-1.412722596433013591e-03,
--2.436052747070789268e-02,
--1.181004380923695793e-02,
--8.055830036755650542e-03,
--6.626879329269285898e-03,
-3.725112725805956870e-02,
-4.241649163886904855e-02,
--1.512466452550143047e-03,
--6.517664128914475719e-02,
--1.133169368695234949e-02,
--3.993257561232894937e-02,
-3.357687836978584354e-03,
--1.214938504272140493e-02,
-8.076140869525261570e-03,
-3.983471600804477930e-03,
-7.495419957558624099e-03,
-1.635326609830372105e-02,
--1.233937148354016279e-02,
-4.849325434770434748e-03,
-1.712186180811841110e-02,
-1.635498097166419168e-02,
-1.929628478246741066e-02,
--3.091572603734675864e-02,
--4.889884497970342419e-04,
-5.382564768660813223e-03,
-2.141074822284281393e-02,
--4.353661211207508763e-02,
-4.526312728412449499e-02,
--1.659549271280411770e-02,
--2.150923216249793837e-02,
-1.693813138408586244e-02,
-1.792303448310121963e-02,
--2.377947850618511438e-02,
--4.007674509193748062e-03,
--2.266779941273853322e-02,
--4.371597495628520499e-03,
-1.137359172298602061e-02,
--4.816858953563496644e-03,
-2.699395010713488002e-02,
--1.732571799584547902e-03,
-2.321885606470459537e-02,
--1.109986796218436281e-02,
--2.294813998742029071e-02,
-8.797803593915886830e-03,
-2.150798275251872976e-02,
--3.055464921053498845e-02,
-1.645020061405375739e-02,
-4.505016483017243098e-03,
-3.597556850290857533e-02,
-2.762623808812349896e-02,
-1.335906013555359165e-02,
-1.773641494874027627e-02,
--1.592383677489124169e-02,
-3.816866672132164517e-02,
-3.667319866362959285e-02,
-8.123216417618549928e-03,
-1.132818537371349575e-04,
-5.785202962579205285e-03,
--9.086049435427411697e-03,
-2.989214300760067913e-03,
--2.320434001740068284e-03,
-1.857377463020384173e-02,
--2.439820137806236744e-02,
--3.227214932267088304e-02,
-5.755101691465825016e-03,
-8.863202817155978880e-03,
-6.657734877080656588e-03,
--3.371864878572523871e-02,
-2.174522746354341438e-02,
--3.688419715035706758e-02,
-9.231698059011250734e-04,
--2.123782858107006039e-02,
--2.080203138757496899e-03,
--5.382566570187918668e-03,
-1.131235545108211227e-02,
-1.057729913154616987e-02,
-4.335442953743040388e-03,
-1.594469393603503791e-03,
--2.356541657354682597e-03,
--1.830369731760583818e-02,
-2.926777087850496037e-05,
--1.745710382238030434e-03,
-1.779290786944329877e-02,
--1.148560601868666743e-02,
--2.998505778377875616e-02,
-5.605372544378042082e-02,
-4.572340706363320524e-03,
-5.513973210327094421e-03,
--1.873485616524703939e-02,
--2.443615157098975171e-02,
--2.645079517969861627e-02,
-6.430327089037746364e-03,
-1.374475716147571802e-03,
--7.686595543054864157e-03,
--3.475821494066622282e-03,
--9.134222217107889827e-03,
-2.807821333117317339e-02,
--1.125976724651991424e-02,
--7.473964471137151469e-03,
-4.789926775265485395e-03,
-3.415293351281434298e-03,
-1.217940150891081395e-02,
-8.009178619249723158e-03,
--2.155545397545211017e-02,
-8.120463046361691903e-03,
--4.170428956858814040e-02,
-7.830668555106968035e-03,
-2.022602204233408044e-02,
--2.028581574559211800e-02,
-1.476033591898158123e-02,
-1.800221301782585145e-02,
--4.244595771888270627e-03,
-3.148810029961168679e-03,
-1.916198317194357445e-02,
--1.155754861720197438e-02,
-2.266841556411236525e-02,
-1.007796088233590091e-02,
--1.379629803705029190e-02,
-5.119774991180747400e-03,
-2.628044958692044020e-02,
--4.968072380870580326e-03,
-6.693437399007962085e-03,
--3.582333787344395976e-02,
--6.908308883430436600e-03,
--7.811704202322289174e-03,
-5.436293828097404912e-03,
-8.152474489179439843e-03,
-5.700851607252843331e-03,
-2.574817868298850867e-03,
-1.521916704368777608e-02,
--1.271394929790403694e-02,
--1.754797146422788431e-02,
-7.316378093673847442e-03,
--9.437051662825979986e-03,
-6.935270212125033140e-03,
-1.245537160430103610e-02,
--1.549466729047708179e-02,
--1.108514906372875045e-02,
-1.883985637687146664e-02,
--2.167562232352793303e-03,
-4.978630872210487376e-03,
-1.247185634711058845e-02,
-4.516550029802601611e-03,
-2.215064737538341259e-03,
-3.784834612160921374e-02,
-3.644732660613954345e-02,
-1.252363299354328739e-02,
--1.599825925222830890e-03,
-7.859462907072156321e-03,
--1.400765036232769559e-02,
--2.645393330603838106e-02,
--3.957919990352820432e-03,
--3.621008343179710123e-03,
-9.219622159434948117e-03,
-7.892735168570651910e-03,
-6.709303286334034402e-03,
-2.628187055233865803e-03,
--2.641390555072575981e-03,
-1.597467749379575391e-02,
--1.410005943616852130e-02,
-2.726363224210217456e-02,
--9.958062891382724741e-03,
-2.021854755003005440e-03,
--1.949160568765364449e-02,
--2.515961613971740007e-02,
-3.628021804615855353e-05,
--4.435380303766578783e-03,
-8.693415485322476127e-03,
--2.752734271343797515e-02,
--1.118746313150040771e-02,
--2.486714554252102832e-02,
--1.740451130899600685e-03,
--2.805474400520324881e-03,
-2.407727323006838421e-02,
-9.672594326548277638e-03,
--2.861085787881165882e-02,
-5.945156694360775526e-03,
--6.579711826052517254e-02,
-4.485669767018407410e-03,
--4.734366791672073652e-03,
--1.752436638344079256e-02,
-2.444782424136065016e-02,
--1.097328063042368740e-02,
--1.356799489876720985e-02,
--5.423655081540345972e-03,
--8.404815156245604504e-03,
-1.221707310178317060e-02,
--2.194811201188713243e-02,
--1.303116338240215520e-02,
-8.578808367019519644e-04,
--1.932932133786380248e-03,
--1.408000102965161274e-02,
--3.274905187427066405e-02,
--1.900599602086004009e-02,
-2.877982434583827714e-02,
--1.410171162424376200e-02,
-5.402604790288023477e-03,
-5.180696793831885034e-03,
--1.497953166544903000e-02,
--3.873560095962602732e-03,
-2.128088058670982857e-02,
--1.413757421541959013e-02,
-2.804182388354092923e-03,
--4.483751393854618419e-03,
-2.057466338155791025e-02,
--4.238008692482253537e-03,
-2.248219889879692415e-02,
-1.664508777321316345e-03,
-1.797259121667593718e-02,
-2.965466707246378151e-03,
-2.934941163402982273e-04,
-2.327450110460631619e-02,
-2.117356623057276011e-02,
--1.407687701866961946e-02,
-1.841561188921332290e-02,
-1.033370214281603694e-02,
--1.366859042173018766e-03,
--1.028503136767540058e-02,
--1.086352934071328422e-02,
--6.701523103401996195e-03,
-3.674290864728391171e-02,
--1.524264742038212647e-02,
--9.899908106308429739e-03,
--8.353408571565523744e-03,
--3.914771694689988743e-03,
--4.490700664537143616e-02,
--5.840281004202552487e-03,
--3.305093582021072229e-02,
-7.584512790344888344e-03,
--6.625544000416994268e-03,
--7.562217180384323348e-03,
-2.297707803663797538e-02,
--1.932510681217536330e-02,
-8.745989389717578888e-02,
-2.840369904879480600e-02,
-1.161251229932531645e-02,
--1.061606977455085109e-02,
-2.267149765975773248e-03,
--4.238370804523583155e-03,
-1.839518087799661000e-02,
-3.598023054692021244e-02,
--1.707434603013098171e-02,
--5.999398666244814930e-03,
--6.438784173224121159e-03,
--2.991240646806545694e-02,
--1.631321643944829564e-02,
-1.330255393695551833e-02,
--3.636382736731320897e-02,
-3.069439905229955773e-02,
-1.363406481686979600e-05,
-2.920672486652620137e-03,
-3.163066445849836011e-02,
-1.760367941460572183e-02,
-5.818970617838204124e-04,
--1.926235171267762919e-02,
-6.500139192212372801e-04,
--3.807854780461639166e-03,
-2.709777031559497049e-02,
-2.156853997439611616e-03,
-3.491104681597789576e-03,
-2.541708177886903503e-04,
--7.745167263783514847e-03,
-3.426024195505306125e-03,
-5.568450629652943679e-04,
--1.578853253740817447e-02,
-4.495091079734266082e-02,
--1.934401383596195956e-02,
--2.090860344469547445e-03,
--5.248717118229251008e-03,
-1.173616184096317773e-02,
--2.040567722578998794e-02,
--5.064112780382856380e-03,
--6.383565504802390760e-03,
--2.526096353540197015e-02,
-4.002936155302450061e-03,
--6.910207923501729410e-02,
--1.071933639468625145e-03,
--1.235872774814197363e-02,
-9.773650366696529090e-03,
--2.149741530884057353e-02,
--3.340818826109170827e-03,
-2.392879425315186381e-02,
--5.437635802081786096e-03,
-1.085583391832187737e-02,
-1.679796114331111312e-02,
-5.936204493627883841e-03,
-2.698491397313773806e-03,
--6.395387928932905631e-04,
-2.538101188256405155e-02,
-2.011224055313505163e-02,
-5.694058863446116794e-03,
--7.966587702685500733e-03,
--1.367631967295892571e-02,
--5.491418783276458097e-03,
--1.030471644480712712e-02,
--5.271261409507133240e-03,
-9.897851571440696022e-03,
-2.710986895181122236e-02,
-2.916913796070730530e-04,
--2.081483757938258281e-02,
--5.033758150602806857e-04,
-1.735652668867260261e-03,
--5.230193650641012590e-03,
--7.767367543419822877e-03,
-2.340998116414993915e-03,
--3.053004795219749104e-04,
-3.388495469698682341e-03,
-1.146385821048170363e-02,
-7.131879683583975380e-05,
--2.212631312431767583e-03,
-3.689170682264375500e-03,
--5.162494222167879690e-03,
--1.359342330833896925e-02,
-2.311418543104082346e-02,
--9.391274248628178611e-03,
--1.762565711978823085e-02,
-6.251880567288026040e-03,
-1.947507957083871588e-03,
-2.795461623463779688e-03,
-1.647698364831740080e-03,
-7.232262112665921450e-03,
-2.688910549040883849e-02,
--4.022772051393985401e-03,
--2.245790499728173018e-03,
-6.667657778598368515e-03,
--1.472272280079778327e-03,
--4.649207085640228235e-03,
--4.195724590681492935e-03,
--5.510665918700396668e-03,
--4.904550116043538202e-03,
--1.096505108580458909e-02,
--1.652644263813272257e-02,
--2.094466497437679278e-02,
-1.490428359829820693e-02,
-3.355966765666380802e-02,
--1.684766332618892193e-02,
--3.217324654106050452e-02,
-2.345243566378485238e-03,
-2.975666499696671963e-02,
--5.894004006404429501e-03,
--5.304902521311305116e-03,
-1.836623017443344066e-02,
--1.041659891052404419e-02,
--1.308841204736381839e-02,
-8.750846795737742684e-03,
--1.585985493147745772e-03,
--2.155492825659166556e-02,
--1.591869792900979519e-02,
-9.432580490829423417e-03,
--1.645074991974979778e-02,
--1.258636700222268666e-02,
--1.627706045983359079e-02,
--5.557938426500185977e-03,
-7.294263719813898206e-03,
--6.796524394303560257e-03,
--6.146338442340493549e-03,
--1.002743665594607592e-03,
--1.544136135489679838e-02,
--1.484100372981629265e-02,
-2.701981835998594830e-02,
-5.035066942218691458e-03,
--1.081432665814645658e-02,
--3.915473644156008959e-03,
-1.116530440631322528e-02,
--7.568762582377530461e-03,
--9.118141054932494063e-03,
-2.363307060295483097e-02,
--9.958223241847008120e-03,
--2.491156398318707804e-02,
-6.787785215419716960e-03,
-2.575042606913484702e-02,
--6.805236090440302914e-03,
-9.676721501455176969e-03,
-5.589777072891592702e-02,
-3.817702585365623019e-04,
--4.339076804462820014e-02,
-3.820459780399687480e-04,
--5.648350075352937330e-03,
--1.472614641534164461e-03,
-1.556033007800579002e-02,
-3.072406929350108770e-02,
--2.922106955156778159e-03,
--1.314932932145893539e-02,
-7.189174648374319348e-05,
--3.909531689714640379e-03,
-9.833761799382046559e-03,
--1.543392202001996408e-02,
--1.873548591975122621e-02,
-1.396073693467769815e-03,
-1.114964298030827160e-02,
--7.766276531037874173e-03,
--9.421445301268250444e-03,
-1.629893528297543526e-02,
-2.271096256736200233e-02,
-1.941514198260847410e-02,
--4.663562576752155748e-05,
--1.059749281557742506e-02,
--1.802444530185311894e-02,
--7.666371113919013061e-03,
--3.442160633858293208e-03,
-1.008582352660596440e-02,
-8.141988539136946201e-03,
-1.397803002037107875e-02,
-4.967673012288287146e-05,
--3.373269668518332626e-02,
--6.694578359019942755e-03,
--1.552248752559535171e-02,
-4.055497207446023822e-03,
--1.449119233657256665e-02,
-3.364010507357306952e-03,
--1.170082582393661098e-03,
--1.833085193356964737e-02,
--1.228844129946082830e-02,
-1.755480601277668168e-02,
-1.129883016692474415e-02,
--1.932738788309507115e-03,
--1.066097709408495540e-02,
--2.368945957277901384e-03,
--1.939881176222115616e-02,
-2.642210719641297986e-02,
-3.523698427015915791e-02,
-2.831683300610165870e-03,
-2.283063313225284358e-02,
-7.182254374492913138e-03,
-7.592860155273229057e-03,
--2.852862208383157711e-02,
--2.290775818109978129e-02,
-4.247987813141662454e-03,
--1.474833546672016382e-02,
-1.022195924306288327e-02,
--3.512530467705800991e-02,
-7.733812279184348881e-03,
--2.876970649231225165e-03,
--5.945696932030841708e-03,
-1.120745067601092160e-02,
-1.347201458411291172e-02,
--5.025910213589668545e-05,
-9.063865797361358306e-03,
--2.138226906536146860e-02,
-5.109271622495725404e-03,
-3.252034211764111976e-02,
-4.813811421627178788e-03,
-1.508184662670828367e-02,
--1.670899503224063803e-02,
-1.853182199411094327e-02,
--4.729678560397587890e-03,
-9.787266026251018047e-04,
--1.496538367355242441e-02,
--2.093779217102564698e-03,
--1.896673912415280895e-03,
--1.081579857855103841e-02,
-3.957933053607120763e-03,
--2.172633457230404089e-02,
--3.457329624798148593e-02,
-1.214827691728714810e-03,
-1.778758710133843124e-02,
-3.549763327755499354e-02,
--4.769720975309610714e-03,
-1.853583410411374643e-02,
-1.140166012919507921e-02,
--1.208045942476019229e-02,
-1.240584195766132256e-02,
-1.667359729763120274e-02,
-5.946851304133815748e-03,
--2.746045141248032277e-03,
--8.850714773871004581e-03,
--3.338876715861261064e-03,
--1.192847864003851975e-02,
-1.199488609127001906e-02,
-1.149435845436528290e-02,
--2.581064609694294631e-02,
--1.658009956590831419e-02,
--2.050778288958099438e-02,
--1.787968560383888100e-02,
-7.742021304147783320e-03,
--1.040473229659255681e-03,
-1.410851093241944840e-02,
-2.733687475847546098e-03,
--3.607173307682387357e-03,
--1.353686344955349306e-02,
--6.597327318741008466e-03,
--2.623887266963720457e-05,
-1.024312098161317408e-02,
-2.999159991741180351e-02,
--2.472464829916134338e-02,
-3.726926597300916823e-03,
--1.232063816860318184e-02,
--9.800773643655702125e-03,
-7.396876498023630293e-03,
--2.098957714333664695e-02,
-2.578735438110015682e-02,
-8.242677577072754502e-03,
-1.317472557420842384e-02,
-1.036536730607622372e-02,
-1.477463480114238337e-02,
-2.154917869338533049e-02,
--1.254921252839267219e-02,
-1.603017965353501634e-02,
--2.382105081342160771e-02,
-1.226091252174228360e-02,
--6.386895756440936003e-03,
--1.730546680919360300e-02,
-7.919914429658092211e-03,
-2.337965881451964378e-03,
-1.062291724374517764e-02,
--1.080543987627606908e-02,
--3.528470997116528365e-03,
-1.129181825090199744e-02,
-1.276588563341647317e-02,
--2.182105662723188115e-03,
--3.664668882265687097e-04,
-1.510079970175865968e-02,
--4.790010943543166269e-03,
--9.372872201493010619e-03,
-7.070297107566148390e-03,
--1.671674638055265037e-03,
-1.543382393429055741e-02,
-1.988150769830099134e-03,
--3.390072542242705649e-03,
-8.580265421187505126e-05,
-1.158945317147299682e-02,
-4.409580267383717347e-03,
-1.178605900495313143e-02,
--8.030706632416695703e-03,
--2.014953478937968681e-02,
--1.642542690387926996e-02,
--1.447836190345697105e-02,
-2.333743995986879011e-02,
-6.274123701587086327e-04,
--1.482019681716337765e-02,
--1.289150690427049925e-02,
-1.551784138428047299e-02,
--5.799001222476363529e-03,
--1.111309549378347583e-02,
-2.339902250096202035e-02,
-1.466969263856299222e-02,
-6.957293767482042139e-03,
--4.126671082340181013e-02,
--1.127030150964856113e-02,
--1.968681471189483861e-02,
-7.790977117838338191e-03,
--6.818551523610949690e-03,
-2.176870656403480028e-02,
-9.962659334996714841e-03,
--2.005878263007616555e-02,
--8.058609580621123314e-02,
--1.115404907613992691e-02,
-1.750798695284174616e-02,
-8.185246033826842557e-03,
--1.044064182788133691e-02,
-1.050373016914818436e-02,
--1.867606198356952562e-03,
-1.164274354814551736e-02,
-1.584685587440617244e-02,
--8.094696007901802659e-03,
-1.618924710637657111e-02,
-1.049668790155919816e-02,
--2.105879418668337236e-02,
-8.109498256817459366e-03,
-1.048479442833922734e-02,
--1.902036584215238647e-03,
-1.951753113535232773e-02,
-6.272327990154735397e-03,
--1.508215380599722301e-02,
--3.108658625569660130e-03,
--2.600858244113624009e-03,
-4.492764989117858584e-03,
-9.510822655283846655e-03,
-2.207101171952672442e-02,
--6.011746393050998102e-03,
-6.285632514118333174e-03,
--1.745754891890101065e-02,
-5.950855642186070500e-03,
-1.085624835395719925e-02,
--1.312971733859740267e-02,
-2.141740457154810498e-02,
-1.625992181361652841e-02,
-1.881551253609359264e-02,
-4.782891842478420412e-03,
-5.902485275873914708e-03,
--8.007584977895021786e-03,
--6.657903955783695199e-04,
--6.338556553123453688e-03,
-1.873688762134406716e-02,
--2.799125295132398519e-03,
--1.485318747327255516e-02,
--1.974958196515217562e-02,
-1.402924317517317832e-02,
-1.600786525523290038e-02,
-1.012879345798864926e-02,
--1.733242431073449621e-02,
-4.443230817560106689e-03,
-1.991378492966760019e-02,
--5.432185268728062337e-03,
-5.151906976243481387e-03,
-1.061795622299541685e-02,
--1.310527623281814227e-02,
-9.549642111960566365e-03,
--1.368443747633136739e-02,
--2.017398658499587258e-02,
--2.980093393489369300e-04,
--1.404808921506628394e-02,
--1.412533291440922802e-02,
--3.294883191119879658e-03,
-4.454243334475904530e-03,
--8.421283273492008448e-03,
-1.694765302818268494e-03,
--1.528159393928945134e-02,
--2.055281335487961838e-02,
-6.169875222258269960e-03,
--2.457632169534917921e-02,
--2.141837933158967713e-02,
-2.076982010694337216e-02,
--1.619636945833917777e-02,
-1.141766506480052998e-02,
-7.418549148133024920e-03,
-1.397676406195387175e-02,
-1.673910088138654689e-02,
-1.613335749425459653e-02,
-5.561829471844248474e-03,
--7.539799564983695919e-03,
--1.220674879732541740e-02,
-3.686573418235639110e-04,
--7.471651983360061075e-03,
--5.863940715789795269e-03,
-6.183539346966426993e-03,
-3.578870330238714901e-03,
-2.367767170071601729e-02,
-2.739032363751903176e-02,
--2.028330962639302015e-02,
-1.344616301357746090e-02,
-1.428120482014492119e-02,
--2.827200508909299883e-03,
--8.189644156482244960e-03,
--2.526918210787698453e-03,
-1.972217550501227309e-02,
-2.116161064186599175e-02,
-1.536526941927149857e-02,
--4.831511533120646726e-03,
-5.996225233684526583e-03,
-3.902642661705613050e-03,
--2.216451888671144843e-03,
--9.701133734779432416e-03,
-6.672447946039028119e-03,
--1.623004267457872629e-02,
-1.459162980318069389e-02,
-2.543707378208637238e-02,
-4.625371028669178659e-03,
--3.959999518701806838e-03,
--1.582958090584725072e-02,
--7.261072139954194427e-03,
-2.129947024368448322e-02,
--8.467121564899570893e-03,
-2.652767434483394027e-04,
--1.617802514228969951e-02,
-1.156479923520237207e-02,
--1.161675553303211962e-02,
-4.582954902434721763e-03,
-2.149930573068559170e-03,
-1.915007895149756223e-03,
-1.154905886651249637e-02,
--1.589805935509502779e-04,
-1.472745466744527186e-02,
-1.787011672277003665e-02,
--1.056974961029482143e-02,
--1.793417420412879437e-02,
-1.266242929116287962e-02,
--9.215665049850940704e-04,
--2.532801583583932370e-03,
--2.488375094253569841e-02,
--6.568782066460698496e-03,
-3.595062880776822431e-05,
--4.463918394321808591e-03,
--6.239975315111223608e-03,
-3.109409038734156552e-03,
-3.349567239056341211e-02,
-3.755253106355667392e-02,
--3.556570957880467003e-03,
--3.110190934967249545e-03,
-2.291007795720361243e-02,
--5.840362713206559767e-03,
-1.526814399985596503e-02,
-1.828139933058992028e-02,
--1.271463426528498565e-02,
-8.613508302005357109e-03,
-9.230293310247361313e-03,
--1.178385390085168270e-03,
-1.423493017064174707e-02,
--1.578779250849038363e-02,
--8.687859072233549904e-03,
-7.870240699048736566e-04,
--7.366210827603936542e-03,
-1.244723338168114354e-02,
--8.318259421503170631e-03,
--1.547621686841011909e-02,
-6.287893085391261030e-03,
-8.703948726179078655e-04,
-3.320142277516424656e-04,
-7.771216175751760429e-03,
--1.346900489879771998e-02,
-3.441873936681076607e-02,
-8.609205285029019467e-03,
--8.592266601044683955e-03,
--1.194944872113410349e-02,
-3.416566182859241685e-02,
--9.320688998559489152e-03,
-1.426526990471757052e-02,
--9.793337713927031257e-03,
--2.002020182553678612e-02,
-3.335717949084937572e-03,
-3.435405265918234363e-03,
--2.699632043950259824e-02,
--1.501216084870975534e-02,
--4.433116503059863697e-03,
-9.655935771297663797e-03,
-9.041206911206245422e-03,
--5.482473351366934540e-03,
-3.876987031253520313e-03,
--9.910093387588858604e-03,
-1.909288071328774138e-03,
-2.029564419353846461e-02,
-2.797715065535157780e-02,
--7.313537393929436976e-03,
--6.261164776515215462e-03,
-1.295709459591307643e-02,
--1.269880364998243793e-03,
-3.818508039694279519e-03,
--8.452152935569756317e-03,
-1.553340954123996233e-02,
-2.976664360612630775e-02,
--2.031826833263039762e-03,
-4.507955517619847974e-02,
--9.827088087331504324e-03,
--1.304875977803021756e-03,
-2.233754518965724856e-02,
-2.232093869242817233e-02,
-4.534311415045522327e-03,
-1.100584900705143740e-02,
--7.354353696064208981e-03,
-2.193159298331011017e-02,
-1.558988518372643808e-02,
--7.747381838271394372e-03,
-2.202932881773449550e-02,
-1.975725939264521072e-02,
--1.170127466320991559e-03,
-2.809884035377763212e-02,
-8.810946630546823144e-03,
-8.596165103517705922e-03,
--1.513510613585822371e-03,
--5.457887674856465657e-03,
-2.454112781379080843e-03,
-5.503082540235482503e-03,
-1.166542526334524155e-03,
--2.770771177019924064e-02,
--3.034130774904042309e-03,
--1.548818042938364739e-02,
-5.488219903782009740e-06,
-9.305982152000069618e-03,
-1.388136615569237654e-02,
--1.501197699690237728e-03,
--3.279524750541895801e-03,
--8.062720626185182482e-03,
-1.290465019410476019e-02,
--1.422986667603254318e-02,
--1.167788947532244406e-03,
-6.230182654689997435e-03,
--4.320573403965681514e-02,
--1.931539920624345680e-02,
-1.881197425827849795e-02,
--9.926602069754153490e-04,
-1.856734289322048426e-02,
--1.000952230242546648e-02,
--1.555628392670769210e-02,
-6.360858789412305106e-03,
-2.275860058143735001e-02,
--2.064108301419764844e-03,
-1.657381430268287797e-02,
-1.888478719629347463e-02,
--1.206091910135000979e-02,
-4.950877034571021622e-03,
--1.063337182276882266e-02,
-2.180212401435710576e-03,
-8.248750364873557589e-03,
-1.766643071314320107e-02,
--2.573822228587232391e-02,
-1.758153943810612102e-02,
--9.062338781586732478e-03,
-5.312807278824038630e-03,
--8.092578715877606771e-03,
-3.541209932882338676e-03,
--1.742221087915822933e-02,
-2.692934195511043072e-02,
-2.778443209826946328e-02,
--3.464140246069291780e-03,
-6.082908064490766611e-03,
-1.153738948050886349e-03,
--8.568021158862394043e-03,
-1.417300651155528612e-02,
-1.703941612504422665e-02,
-9.536510368343442332e-03,
--1.912544893275480674e-03,
-2.305097121279686689e-03,
--2.774106542346999049e-03,
--2.642214007209986231e-04,
-3.488037901115603832e-03,
--6.083788762043696070e-03,
-7.631018670508638024e-03,
-6.208615624927915633e-03,
--1.816763923707185452e-02,
--3.551857552884030069e-03,
-1.038769725710153580e-03,
--9.091123152757063508e-03,
-1.276774746511364321e-02,
--7.004351704381406090e-04,
--5.962423281744122072e-04,
--3.703397407662123333e-04,
-7.641337983659468765e-03,
-2.405025323969312012e-02,
-8.283065244904718946e-03,
--6.581717793596908624e-03,
-2.355621678289026091e-02,
-5.711935624276520684e-03,
--1.963596163550391724e-02,
--8.253962613525799713e-03,
--7.611531538350391549e-03,
--1.166138745102216455e-02,
-1.406958542356733302e-02,
--8.267057477496564171e-04,
--1.614632323980913454e-03,
-1.037310075189452549e-02,
--1.308949106896761805e-02,
--7.686795340850949634e-03,
--1.066315966309048201e-02,
-6.507713886094279587e-03,
--3.672932118934113706e-03,
-9.188450873989496143e-03,
--3.927565371122909363e-03,
-5.025976331671700228e-03,
-1.134529893752187456e-02,
-4.721692262683064092e-03,
--2.841148190200328688e-02,
--2.588801622623577645e-02,
-1.385693862685002421e-02,
-5.379268361139111221e-03,
-8.438213402405381550e-03,
-4.820020956685766578e-03,
-9.436098279547878304e-03,
--1.577339513460174203e-03,
-9.547394874971359697e-05,
-8.792690816335380077e-03,
-1.876135047568823142e-02,
--1.338109532371163299e-02,
--1.352524380490649492e-02,
--1.785166300833225181e-02,
--2.956867759348824447e-03,
-1.104491936857812141e-02,
--1.544391745701432193e-02,
-3.505717474035918799e-03,
-5.319875470013358046e-03,
-5.349700874648988420e-03,
--5.243656209131586365e-03,
-6.656189449131489017e-04,
-1.047522887820378015e-02,
-3.999249893240630627e-03,
--7.029033807339146560e-03,
--2.333188551710918632e-03,
--9.972017270047216914e-03,
--1.209601759910583496e-04,
-1.787121039742487622e-02,
--2.332056751474738052e-02,
--8.481634623603896000e-03,
--2.115292713278904644e-03,
--4.878579199521481877e-03,
-2.901614044094458396e-03,
-1.118667477858252893e-02,
--1.643372207763604820e-02,
--3.780025200103409762e-03,
-1.021935343160293962e-02,
--1.663857762468978719e-03,
-1.278010138776153326e-03,
-1.340722729510162095e-02,
-5.887076932413038040e-03,
-3.500899195205420429e-03,
--4.137214517686516220e-03,
--4.135751244029961179e-03,
--1.496108922874555061e-02,
--7.423543953336775303e-03,
-6.061168492306023663e-03,
-2.628208097303286021e-03,
--2.138131599640473793e-02,
-1.983559717773459782e-02,
--7.231167698046192180e-03,
-2.858428512990940189e-03,
--1.807744955876842280e-02,
--3.817988609080202964e-03,
-1.027301762253046008e-01,
--1.049182281713001517e-03,
--8.274902240373194218e-03,
-2.842782252810138570e-03,
-8.230978345090988677e-03,
--1.606798908323980911e-03,
-2.794570516562089237e-03,
--3.883002186194062320e-03,
-3.851585538359359080e-03,
-6.174380355514585798e-03,
--9.910328217665664505e-03,
--4.128728516661794856e-03,
--5.869911052286624908e-03,
-3.374529173015616834e-03,
-1.397572822752408614e-02,
--8.214152156142517924e-03,
--1.548986030975356767e-02,
--1.815933951293118367e-02,
--1.041212175914552063e-02,
-7.914100983180105339e-03,
-5.982818244956433600e-03,
--2.118119391961954465e-02,
-7.956661866046488285e-04,
-2.486425449023954744e-03,
-7.729447114979848092e-03,
--2.643775776959955692e-02,
--6.299247796414419662e-03,
-1.569755345117300818e-02,
-2.231216741347452626e-02,
-2.308206413872539858e-02,
--6.817292392952367880e-04,
-2.169419076744816353e-03,
--6.020447395712835714e-03,
--1.144130285538267310e-02,
--1.785701161716133495e-02,
--7.975140429334715703e-03,
--1.460391271893968224e-02,
--1.315491588611621479e-02,
--1.715717080514878102e-05,
--2.053622798412107001e-02,
-8.538523814058862973e-03,
-1.342314690118655632e-02,
-1.024173995892851963e-02,
--4.919437650823965495e-03,
-4.940598583198152136e-03,
--3.211722841951995810e-02,
--1.490932577289640938e-02,
--4.623951171379303386e-03,
-2.821497619152069092e-03,
--2.649463398847729054e-03,
--1.405036709038540686e-02,
-1.470845625153742800e-02,
-8.838723394364932789e-03,
--1.557127897103782700e-03,
--9.210367813648191043e-03,
-5.304105958202854016e-04,
--2.531161342631094158e-03,
--1.434937049634754701e-03,
-3.703486156882718205e-02,
-1.880257999291643639e-02,
--1.643821003381162925e-03,
--7.333502569235861128e-03,
-9.127936000004411091e-03,
--7.647617696784436356e-03,
-6.739112496143206607e-03,
--7.894508977187797644e-04,
--1.318564581597456674e-02,
--1.248691959772258928e-02,
-6.605757673969492490e-03,
-4.590558167546987360e-03,
--6.306250725174322724e-03,
-7.443017896730452602e-03,
-2.741795399924740022e-03,
--8.752206584904342196e-03,
--7.495709529030137111e-03,
-1.216502844472415815e-02,
--4.111754623590968376e-03,
-4.453295689017977457e-04,
--1.709070159122347971e-02,
-1.102580450242385268e-02,
-2.065254080516751994e-02,
--8.239158484502694807e-03,
-9.233326080720870993e-03,
-1.985971657559275697e-02,
--2.395824002451263542e-03,
-4.743019084708066310e-03,
-6.525630661781178614e-03,
--2.065923289046622813e-02,
--1.649279161938466132e-03,
--1.783575213630683720e-02,
-4.754915452213026757e-03,
-9.640499942906898684e-03,
-1.249273929279297660e-02,
-2.034188253455795209e-02,
--5.028525445959530424e-03,
--3.322070196736603889e-03,
--1.645722325192764540e-02,
-6.581127032404765704e-04,
-7.320888177491724665e-03,
-7.856722714495844445e-03,
-4.388645774452015584e-03,
--4.550904226198327143e-03,
--4.750136411166750242e-03,
-8.064571651630103588e-03,
-1.402271989500150119e-02,
--1.569007054786197841e-02,
--1.265118458541110089e-02,
--7.055002311244606798e-03,
-6.481654668459668582e-03,
-6.615672711632214487e-03,
-2.189354971051216125e-02,
-5.599562628776766533e-03,
-3.221723822643980661e-02,
--1.335754067185917007e-02,
-2.201184382771543754e-02,
--7.557125013408949772e-03,
--5.908043868839740580e-03,
--6.321127594856079494e-03,
--1.114656733989249915e-02,
--3.561108367284759792e-03,
-2.480515344068407943e-02,
--1.113502854714170168e-02,
-1.354382369900122320e-02,
-1.089393672300502727e-02,
-3.519720117037650024e-03,
--8.801410032901913361e-04,
-3.806073726445902226e-03,
-2.823798181489110132e-02,
-3.081401831004768749e-02,
--2.284305938519537622e-03,
-1.698123157257214145e-03,
--9.708647181105334659e-03,
-5.234614625805988908e-03,
--4.274043971963692630e-03,
-1.920406317804008653e-02,
-2.079129822668619385e-02,
--5.977447413670233800e-03,
--1.238779033010359819e-02,
--6.467461201827973474e-03,
--1.018362818285822868e-04,
-3.551157281617633825e-03,
--5.910143864821293218e-03,
--5.956706445431336761e-05,
--1.121888507623225451e-02,
-8.594607788836583495e-03,
-2.071786423912271718e-02,
-4.774379002628848130e-03,
-1.278833637479692784e-03,
--1.634748526412295108e-02,
--1.983580389060080051e-02,
--6.526824960019439459e-03,
-1.258983963634818836e-03,
--1.009841045597568120e-02,
-6.983118969947100033e-03,
--1.094209363654954394e-02,
-5.126653431216254657e-03,
--2.888724498916417360e-03,
--1.769471401348710017e-03,
-2.839042840641923388e-02,
--3.108765103388577526e-03,
-6.882070490974001586e-03,
-7.991266949102283218e-03,
--8.260777910800242915e-03,
-1.065625900264421944e-02,
-3.254883131012320518e-02,
-1.444831169646931856e-03,
--1.047802006360143423e-02,
--2.441555096447700721e-03,
--2.320154837070731359e-03,
--1.131398851866833796e-02,
-6.709687170223332595e-03,
-8.056852873414755214e-03,
--1.638028165325522510e-03,
--6.307428656145930464e-03,
-7.085928921151207413e-03,
--6.061694107484072273e-03,
-1.377297842191183095e-03,
-2.058515690441709012e-02,
--7.517418754287063902e-03,
--6.791737818275578320e-03,
--9.608150122221558592e-03,
--1.765466118667973275e-02,
-1.171345178736373817e-02,
--7.566322710772510618e-04,
--2.569345899391919236e-02,
-8.358489390229806998e-03,
-4.053028940688818693e-03,
--7.538855471648275679e-03,
--5.413519719149917017e-03,
--8.694138034479692231e-03,
-2.115376177243888378e-03,
--1.046781857730820844e-02,
-1.387932138750329683e-02,
-1.354039473408192792e-02,
--1.206661026226356695e-02,
-2.293357864255085588e-02,
--3.610883929650299078e-03,
--1.243656184815336053e-02,
-4.586090275552123785e-03,
-5.200736594269983123e-03,
-7.774297015566844146e-03,
--1.043932152679189970e-02,
-1.743367178132757683e-02,
-8.978349773678928267e-03,
-2.936053164303302834e-02,
--1.515399164054542693e-03,
-1.559907675255089998e-02,
-2.093702155398204784e-02,
--2.587611760827712541e-03,
-9.490785218076780438e-03,
-1.903583339299075305e-02,
--9.700666961725801576e-03,
-1.582571490143891467e-03,
--1.267328932590316894e-03,
-2.362947221263311717e-03,
-1.946196645149029722e-02,
-9.011766636103858003e-03,
-9.982516494346782565e-03,
--1.225340649543795769e-02,
--9.711418789811432709e-03,
--1.911661276244558307e-03,
--7.381553962477483029e-03,
--8.592037385824369700e-04,
--2.242713966406881948e-02,
--1.917882619891315629e-02,
--1.247785883024334942e-02,
-1.914953821687959048e-03,
--8.775244912249036480e-03,
-3.990936730406247253e-03,
--3.856583044398576195e-03,
--2.273750600579660386e-03,
--1.436085702443961123e-02,
--4.667493810120504523e-03,
--4.294897755698912079e-04,
-2.130453842692077090e-02,
--8.692576014436782100e-03,
--1.817197035998106142e-02,
--6.492670497391372746e-03,
--7.498866040259599339e-03,
--5.412057670127978998e-04,
--1.215554334339685695e-03,
--3.073696017963811755e-03,
-1.334991632029414264e-03,
--3.550659447500948036e-03,
--9.801604400854558100e-03,
-2.646082234568893771e-02,
-1.637801051838323405e-02,
--1.874335035681724479e-02,
--8.882203121902420423e-03,
-1.042715916992165148e-02,
--8.008928503841161945e-04,
-3.445437260961625552e-03,
--2.066332043614238674e-03,
--1.658553721281350729e-02,
--4.701998946984531358e-03,
--2.336643333546817129e-03,
-5.569492704307776702e-03,
--6.626862176926806396e-03,
-6.663641403429209839e-03,
-2.671076881233602675e-03,
-4.870861949166282664e-03,
-2.109902994707226614e-02,
-2.374765824060887098e-02,
--1.163802237570052932e-02,
-2.233360558748245100e-02,
-3.538694733288139192e-03,
--2.363352337852120345e-04,
-1.954562460596207668e-02,
--5.183506470348220492e-03,
-5.724921787623316397e-03,
-8.219848998123779893e-03,
-6.348441937007009549e-04,
--5.559814680600539197e-03,
-1.281951499549904838e-02,
--1.220506809186190297e-02,
-1.891820076853036742e-02,
--8.371846610680222164e-03,
--1.943286275491118431e-03,
-3.899429836019408045e-03,
-1.576097751240013167e-03,
-7.976570978644303284e-03,
-4.187518160324543888e-03,
-6.003536604112014174e-03,
--9.083423356059938669e-03,
-1.936828077305108309e-02,
-2.860336349112912839e-03,
--1.972673910204321077e-02,
-1.472707611828809673e-02,
--3.324280976812588159e-04,
--2.488546670065261424e-03,
--5.497445724904537548e-03,
-1.623212448321282794e-02,
-2.072620307444594800e-02,
--1.351172099821269443e-02,
-1.657902354490943184e-02,
-8.916535173193551253e-03,
--5.132115678861737251e-03,
--1.732679416454629973e-02,
--1.084842742420732888e-03,
--4.556570958811789643e-04,
--1.681491440685931732e-02,
--8.767609047936275601e-03,
--4.238943082600599328e-03,
--1.592217848738073510e-02,
-7.929830462671816349e-03,
-8.896900151739828289e-03,
--4.933272162452340473e-03,
-1.622654025995871191e-02,
--6.320005346788093101e-03,
--4.353146724861289493e-03,
-6.533471448346972292e-03,
--4.855240543838590730e-03,
-2.668016718234867031e-03,
-1.574495651530014723e-02,
--6.509712709521409471e-03,
--5.505113236722536385e-03,
-2.978651813464239240e-03,
-1.452154327853349969e-02,
-1.062976232788059860e-02,
-1.576894427998922626e-02,
-2.189539915416389634e-02,
--3.995549643877893860e-03,
-3.494401695206761577e-04,
--6.171266562887467444e-03,
-1.223825230845250120e-02,
--2.051482626702636619e-02,
--6.103948038071393793e-03,
-1.262372150813462213e-02,
-2.649483201093971729e-02,
-2.597295416053384612e-02,
-5.615037761162966663e-03,
-1.238398958230391231e-03,
--5.837172879546414714e-03,
--3.033868223428726370e-03,
-2.823848689149599696e-03,
--4.485397596727125520e-03,
-8.510926156304776669e-03,
-3.129494094755500694e-02,
-3.058735091326525261e-03,
-7.470455323345959360e-03,
-3.947153632179833775e-03,
-1.762350695207715034e-02,
-3.305345369153655892e-03,
--9.971052647051692822e-03,
-2.668607260129647260e-02,
-2.279791150358505616e-02,
-1.376637805224163388e-02,
--3.190379391890019265e-03,
--1.087475954118417507e-02,
--5.232337198685854326e-03,
--1.684350024443119834e-02,
--1.514592983294278314e-02,
--8.012250671163201679e-03,
-5.143893556669354786e-03,
-8.547383337281643997e-03,
-1.025208962091710488e-02,
-2.076280133333057235e-02,
--3.050260457675903897e-02,
--3.929960570530965924e-03,
-7.651075691683217883e-03,
--6.275656138313934424e-04,
-8.716393106442409805e-03,
-6.821079288056352934e-03,
--5.091704278311226343e-03,
-1.078921162988990580e-02,
-3.676004935987293998e-02,
--8.612308790907264103e-03,
--7.157141552306711327e-03,
--6.438640132546425039e-03,
--3.417930004070513037e-03,
--1.059016523067839513e-02,
-3.383215147187001916e-03,
--1.310363429365679665e-03,
--1.758118880534311618e-02,
-2.436707203742116744e-04,
--1.667125566746108147e-03,
--1.304479663156143943e-02,
--1.887810298940166900e-02,
-4.702559587894938713e-03,
--9.527037036605179310e-04,
-1.051573613367509132e-02,
--1.535475985147058929e-02,
-8.621049416251479885e-03,
--7.776543556246906519e-03,
-1.011012516682967613e-02,
-1.165708121552597697e-02,
--4.989801449846708800e-03,
--8.666561171412467263e-03,
-1.671402715146541595e-04,
--3.973259581835008811e-03,
-8.744894462870434274e-03,
--8.055317634716629288e-03,
--7.455401920014992530e-03,
--1.751273257832508534e-02,
-1.859033247019397164e-02,
--2.095822060364298592e-02,
-2.669851644895970821e-02,
--1.979418627452105353e-02,
--3.342758317012339917e-03,
-1.506193616660311861e-03,
-1.006398356985300710e-02,
--1.015973714675055817e-02,
-1.706497468985617230e-02,
-2.065443931496702135e-02,
-3.853589061181992431e-02,
--6.140403423341922289e-03,
-4.162678774446248835e-03,
-4.828265311516588445e-04,
-3.613275947063811916e-02,
-1.099346728879027103e-02,
--3.367625118698924997e-04,
-2.904278510250151227e-02,
-4.450384952360763972e-03,
--6.147134120692498808e-03,
--1.189552596188150381e-02,
--2.196746382978744666e-02,
-1.142448401078581879e-02,
-4.922766225645318350e-02,
--1.422957038885215228e-02,
-2.952824942767619948e-02,
--3.609892074018716812e-02,
-2.216266572213498875e-02,
--3.879444045014679432e-02,
--7.506611669668927789e-03,
-3.744107810198329250e-02,
-1.787485381428268599e-02,
-5.358831491321325519e-04,
--2.952785363886505435e-02,
--1.241250551538541913e-02,
-7.306755555327981363e-03,
-1.529467361979186492e-03,
--4.716435603040735995e-03,
-1.277329750591888954e-02,
-2.542682273779064278e-02,
-1.427434660727158224e-02,
--5.086504568462259929e-02,
--1.143796327523887227e-02,
-2.853804218466393650e-02,
-7.917443204496521850e-03,
--2.364957803802099010e-02,
--6.276517850346862923e-03,
-1.087480059941299190e-02,
-1.766669069183990359e-02,
--1.478109410381875898e-02,
--1.411852211458608551e-03,
-2.858506352640688593e-03,
-3.967653827276081124e-02,
-2.291722809895873139e-02,
-2.009000665930216287e-03,
-9.199097063537919547e-03,
--6.308554112911224712e-03,
--9.577488897775765075e-04,
--2.451680449303239497e-03,
--2.544487451668828865e-02,
--6.693482377158943564e-03,
--1.901679844886530052e-02,
-2.065804268931970000e-02,
-5.487208795966580337e-03,
--6.419890472898259759e-03,
-9.942441206658259728e-03,
-3.646264501148834686e-02,
--2.859469313407316665e-02,
-1.080586735624819959e-02,
-7.148144890379626304e-03,
--1.689901633653789759e-03,
--4.263541516847908774e-02,
-3.498093265807256276e-03,
--4.083726386306807052e-03,
-2.253368667443282972e-02,
-2.504280641733203083e-02,
-2.352550349314697087e-02,
--7.661223749164491718e-03,
--1.995733280273270793e-02,
--2.141489349887706201e-02,
--2.271508588455617428e-02,
--3.110957400058396091e-02,
--7.970536476932466247e-04,
--2.135029069031588872e-02,
--1.615363174350932171e-02,
--7.376620118157007147e-03,
-1.136203194691915927e-02,
--2.649353645974770327e-02,
-2.765877211495535587e-02,
-2.074462210293859157e-03,
-2.204873053124174406e-02,
--1.883032270707190106e-02,
--1.732160617830231716e-02,
-2.236614992725662748e-02,
-9.370663436129688783e-04,
-2.762139044934883853e-02,
--1.552296521549578756e-02,
-1.774471122771501680e-02,
--1.831492233322933386e-02,
-1.848664225253742258e-02,
-1.133294749888591468e-02,
--1.446236744523048470e-02,
--3.690123688429593901e-02,
--3.819424484390765429e-02,
-1.003097087959758937e-02,
-2.435783122782595594e-02,
--2.381120717618614296e-02,
--2.809699791250750561e-02,
-2.886609745910391142e-02,
-7.066190676414407963e-03,
--1.016267060185782611e-02,
-2.805423315148800750e-02,
--1.219403161230729962e-02,
-4.133120295591652393e-03,
--1.520158144994638791e-02,
--1.997057115659117699e-04,
--2.110574464313686067e-03,
-2.240938654867932062e-02,
--5.625425074249505719e-02,
-5.199041659943759094e-03,
--1.721694635052699665e-02,
--7.853147701825945876e-03,
--7.290923979599028999e-03,
--2.261036707204766572e-02,
--1.738484574307221781e-02,
-2.255194803583435853e-02,
-8.489550091326236725e-04,
--5.013549267459893599e-03,
--6.900661801046226521e-03,
--2.434362768835853646e-02,
--2.084533177549019282e-03,
--1.376168632123153657e-02,
-1.211340274894610013e-02,
--3.331455499865114411e-02,
-2.792396319564431975e-02,
-2.727972451393725212e-02,
--1.137276639929041180e-03,
-2.107373339822515845e-02,
-2.057588454335928033e-02,
-1.830648991162888567e-02,
--2.273723261896521027e-03,
--1.929803152161184768e-02,
-1.351142524072201923e-02,
-4.411324090324342077e-03,
--1.704691734630614519e-03,
--6.175023515243083412e-03,
-1.413247554155532340e-02,
--3.042460333090275457e-02,
--4.074956616386770855e-03,
--1.725879134610295434e-02,
--2.190176274161785910e-02,
-1.397514635464176466e-02,
-1.157898288602154935e-02,
-1.116817494003044010e-02,
-7.384492515120655645e-03,
-1.418961353047052364e-02,
--2.963992182776564732e-02,
--2.093353134114295244e-02,
--2.037936307824566098e-02,
--1.645624818047508597e-02,
-3.182987581240013381e-02,
--1.223997086053714087e-02,
-2.495519933290779591e-03,
--5.320763571653515378e-02,
-4.535343381576240063e-03,
--2.041436027502641082e-02,
--1.944463358377106557e-02,
--2.277987955603748629e-02,
-1.928524321992881527e-02,
--1.183865525817964261e-02,
-4.891169200345757244e-03,
--1.773684252984821658e-02,
--1.869004432810470601e-02,
--6.452457618433981960e-03,
-1.498907190398313062e-03,
-3.343018507584929327e-02,
-5.797518910840153417e-02,
--1.128418180560402176e-02,
-1.457973863289225777e-02,
--3.033333667553961364e-03,
--4.012424062239006162e-03,
--1.148171606473624733e-04,
--9.845346826477908292e-03,
-5.389356467640027057e-03,
--1.688954794080927899e-02,
--3.054580911993980477e-02,
-2.922235441801603745e-02,
-1.872595368186012049e-02,
--4.471853582072071558e-03,
-5.528640432748943914e-03,
--1.966052072239108522e-02,
-2.040305837581399967e-02,
--8.859847276471554539e-03,
--2.194737640093080699e-02,
-1.203827609715517680e-02,
--2.528259065002202852e-05,
-3.109157728031277518e-02,
-1.506878216750919854e-02,
--4.009108245372772217e-02,
--1.503079550166148623e-02,
--3.743553836829960346e-02,
--3.348496071994304379e-02,
--1.156956416089087794e-02,
-9.587417333386838089e-03,
--3.091484878823393817e-02,
--9.297358104959130287e-03,
--6.026116339489817966e-03,
-2.803363663260825117e-02,
-1.710467734956182609e-02,
--2.473799009458161821e-02,
-3.801211167592555371e-03,
--3.406440973049029534e-03,
-2.679838360054418583e-02,
-2.099937659222632716e-02,
-2.547775104176252972e-04,
--2.136118641356006195e-02,
-4.712846358306706229e-02,
-5.675952713936567584e-02,
-1.435134616913274000e-02,
--9.094646415906026266e-03,
-2.013558214530348708e-02,
-3.569096239516511559e-02,
--4.260482303798198561e-02,
-1.622531267057638477e-02,
--1.064220830448903143e-02,
-4.038759353570640226e-02,
-2.005523220868781278e-02,
-1.869875285774469419e-03,
--1.200556157709797844e-02,
--8.927008265163748763e-03,
-1.102921260753646562e-02,
--1.141999643296003411e-02,
--6.043970497557893260e-05,
--1.000145886064274237e-02,
--2.781596259446814795e-02,
--2.155645787715911726e-02,
-5.072970921173691576e-03,
-1.027262099378276582e-02,
-1.866447147331200609e-02,
--3.279702056897803968e-03,
-1.524280269513838063e-02,
--2.853577626228798354e-03,
--1.595507248712237786e-02,
--2.190637937746942217e-03,
-1.666411121841520140e-02,
-1.899441269924864034e-02,
-4.048163760453462462e-02,
-1.412722596433013591e-03,
--2.436052747070789268e-02,
--1.181004380923695793e-02,
--8.055830036755650542e-03,
--6.626879329269285898e-03,
-3.725112725805956870e-02,
-4.241649163886904855e-02,
--1.512466452550143047e-03,
--6.517664128914475719e-02,
--1.133169368695234949e-02,
--3.993257561232894937e-02,
-3.357687836978584354e-03,
--1.214938504272140493e-02,
-8.076140869525261570e-03,
-3.983471600804477930e-03,
-7.495419957558624099e-03,
-1.635326609830372105e-02,
--1.233937148354016279e-02,
-4.849325434770434748e-03,
-1.712186180811841110e-02,
-1.635498097166419168e-02,
-1.929628478246741066e-02,
--3.091572603734675864e-02,
--4.889884497970342419e-04,
-5.382564768660813223e-03,
-2.141074822284281393e-02,
--4.353661211207508763e-02,
-4.526312728412449499e-02,
--1.659549271280411770e-02,
--2.150923216249793837e-02,
-1.693813138408586244e-02,
-1.792303448310121963e-02,
--2.377947850618511438e-02,
--4.007674509193748062e-03,
--2.266779941273853322e-02,
--4.371597495628520499e-03,
-1.137359172298602061e-02,
--4.816858953563496644e-03,
-2.699395010713488002e-02,
--1.732571799584547902e-03,
-2.321885606470459537e-02,
--1.109986796218436281e-02,
--2.294813998742029071e-02,
-8.797803593915886830e-03,
-2.150798275251872976e-02,
--3.055464921053498845e-02,
-1.645020061405375739e-02,
-4.505016483017243098e-03,
-3.597556850290857533e-02,
-2.762623808812349896e-02,
-1.335906013555359165e-02,
-1.773641494874027627e-02,
--1.592383677489124169e-02,
-3.816866672132164517e-02,
-3.667319866362959285e-02,
-8.123216417618549928e-03,
-1.132818537371349575e-04,
-5.785202962579205285e-03,
--9.086049435427411697e-03,
-2.989214300760067913e-03,
--2.320434001740068284e-03,
-1.857377463020384173e-02,
--2.439820137806236744e-02,
--3.227214932267088304e-02,
-5.755101691465825016e-03,
-8.863202817155978880e-03,
-6.657734877080656588e-03,
--3.371864878572523871e-02,
-2.174522746354341438e-02,
--3.688419715035706758e-02,
-9.231698059011250734e-04,
--2.123782858107006039e-02,
--2.080203138757496899e-03,
--5.382566570187918668e-03,
-1.131235545108211227e-02,
-1.057729913154616987e-02,
-4.335442953743040388e-03,
-1.594469393603503791e-03,
--2.356541657354682597e-03,
--1.830369731760583818e-02,
-2.926777087850496037e-05,
--1.745710382238030434e-03,
-1.779290786944329877e-02,
--1.148560601868666743e-02,
--2.998505778377875616e-02,
-5.605372544378042082e-02,
-4.572340706363320524e-03,
-5.513973210327094421e-03,
--1.873485616524703939e-02,
--2.443615157098975171e-02,
--2.645079517969861627e-02,
-6.430327089037746364e-03,
-1.374475716147571802e-03,
--7.686595543054864157e-03,
--3.475821494066622282e-03,
--9.134222217107889827e-03,
-2.807821333117317339e-02,
--1.125976724651991424e-02,
--7.473964471137151469e-03,
-4.789926775265485395e-03,
-3.415293351281434298e-03,
-1.217940150891081395e-02,
-8.009178619249723158e-03,
--2.155545397545211017e-02,
-8.120463046361691903e-03,
--4.170428956858814040e-02,
-7.830668555106968035e-03,
-2.022602204233408044e-02,
--2.028581574559211800e-02,
-1.476033591898158123e-02,
-1.800221301782585145e-02,
--4.244595771888270627e-03,
-3.148810029961168679e-03,
-1.916198317194357445e-02,
--1.155754861720197438e-02,
-2.266841556411236525e-02,
-1.007796088233590091e-02,
--1.379629803705029190e-02,
-5.119774991180747400e-03,
-2.628044958692044020e-02,
--4.968072380870580326e-03,
-6.693437399007962085e-03,
--3.582333787344395976e-02,
--6.908308883430436600e-03,
--7.811704202322289174e-03,
-5.436293828097404912e-03,
-8.152474489179439843e-03,
-5.700851607252843331e-03,
-2.574817868298850867e-03,
-1.521916704368777608e-02,
--1.271394929790403694e-02,
--1.754797146422788431e-02,
-7.316378093673847442e-03,
--9.437051662825979986e-03,
-6.935270212125033140e-03,
-1.245537160430103610e-02,
--1.549466729047708179e-02,
--1.108514906372875045e-02,
-1.883985637687146664e-02,
--2.167562232352793303e-03,
-4.978630872210487376e-03,
-1.247185634711058845e-02,
-4.516550029802601611e-03,
-2.215064737538341259e-03,
-3.784834612160921374e-02,
-3.644732660613954345e-02,
-1.252363299354328739e-02,
--1.599825925222830890e-03,
-7.859462907072156321e-03,
--1.400765036232769559e-02,
--2.645393330603838106e-02,
--3.957919990352820432e-03,
--3.621008343179710123e-03,
-9.219622159434948117e-03,
-7.892735168570651910e-03,
-6.709303286334034402e-03,
-2.628187055233865803e-03,
--2.641390555072575981e-03,
-1.597467749379575391e-02,
--1.410005943616852130e-02,
-2.726363224210217456e-02,
--9.958062891382724741e-03,
-2.021854755003005440e-03,
--1.949160568765364449e-02,
--2.515961613971740007e-02,
-3.628021804615855353e-05,
--4.435380303766578783e-03,
-8.693415485322476127e-03,
--2.752734271343797515e-02,
--1.118746313150040771e-02,
--2.486714554252102832e-02,
--1.740451130899600685e-03,
--2.805474400520324881e-03,
-2.407727323006838421e-02,
-9.672594326548277638e-03,
--2.861085787881165882e-02,
-5.945156694360775526e-03,
--6.579711826052517254e-02,
-4.485669767018407410e-03,
--4.734366791672073652e-03,
--1.752436638344079256e-02,
-2.444782424136065016e-02,
--1.097328063042368740e-02,
--1.356799489876720985e-02,
--5.423655081540345972e-03,
--8.404815156245604504e-03,
-1.221707310178317060e-02,
--2.194811201188713243e-02,
--1.303116338240215520e-02,
-8.578808367019519644e-04,
--1.932932133786380248e-03,
--1.408000102965161274e-02,
--3.274905187427066405e-02,
--1.900599602086004009e-02,
-2.877982434583827714e-02,
--1.410171162424376200e-02,
-5.402604790288023477e-03,
-5.180696793831885034e-03,
--1.497953166544903000e-02,
--3.873560095962602732e-03,
-2.128088058670982857e-02,
--1.413757421541959013e-02,
-2.804182388354092923e-03,
--4.483751393854618419e-03,
-2.057466338155791025e-02,
--4.238008692482253537e-03,
-2.248219889879692415e-02,
-1.664508777321316345e-03,
-1.797259121667593718e-02,
-2.965466707246378151e-03,
-2.934941163402982273e-04,
-2.327450110460631619e-02,
-2.117356623057276011e-02,
--1.407687701866961946e-02,
-1.841561188921332290e-02,
-1.033370214281603694e-02,
--1.366859042173018766e-03,
--1.028503136767540058e-02,
--1.086352934071328422e-02,
--6.701523103401996195e-03,
-3.674290864728391171e-02,
--1.524264742038212647e-02,
--9.899908106308429739e-03,
--8.353408571565523744e-03,
--3.914771694689988743e-03,
--4.490700664537143616e-02,
--5.840281004202552487e-03,
--3.305093582021072229e-02,
-7.584512790344888344e-03,
--6.625544000416994268e-03,
--7.562217180384323348e-03,
-2.297707803663797538e-02,
--1.932510681217536330e-02,
-8.745989389717578888e-02,
-2.840369904879480600e-02,
-1.161251229932531645e-02,
--1.061606977455085109e-02,
-2.267149765975773248e-03,
--4.238370804523583155e-03,
-1.839518087799661000e-02,
-3.598023054692021244e-02,
--1.707434603013098171e-02,
--5.999398666244814930e-03,
--6.438784173224121159e-03,
--2.991240646806545694e-02,
--1.631321643944829564e-02,
-1.330255393695551833e-02,
--3.636382736731320897e-02,
-3.069439905229955773e-02,
-1.363406481686979600e-05,
-2.920672486652620137e-03,
-3.163066445849836011e-02,
-1.760367941460572183e-02,
-5.818970617838204124e-04,
--1.926235171267762919e-02,
-6.500139192212372801e-04,
--3.807854780461639166e-03,
-2.709777031559497049e-02,
-2.156853997439611616e-03,
-3.491104681597789576e-03,
-2.541708177886903503e-04,
--7.745167263783514847e-03,
-3.426024195505306125e-03,
-5.568450629652943679e-04,
--1.578853253740817447e-02,
-4.495091079734266082e-02,
--1.934401383596195956e-02,
--2.090860344469547445e-03,
--5.248717118229251008e-03,
-1.173616184096317773e-02,
--2.040567722578998794e-02,
--5.064112780382856380e-03,
--6.383565504802390760e-03,
--2.526096353540197015e-02,
-4.002936155302450061e-03,
--6.910207923501729410e-02,
--1.071933639468625145e-03,
--1.235872774814197363e-02,
-9.773650366696529090e-03,
--2.149741530884057353e-02,
--3.340818826109170827e-03,
-2.392879425315186381e-02,
--5.437635802081786096e-03,
-1.085583391832187737e-02,
-1.679796114331111312e-02,
-5.936204493627883841e-03,
-2.698491397313773806e-03,
--6.395387928932905631e-04,
-2.538101188256405155e-02,
-2.011224055313505163e-02,
-5.694058863446116794e-03,
--7.966587702685500733e-03,
--1.367631967295892571e-02,
--5.491418783276458097e-03,
--1.030471644480712712e-02,
--5.271261409507133240e-03,
-9.897851571440696022e-03,
-2.710986895181122236e-02,
-2.916913796070730530e-04,
--2.081483757938258281e-02,
--5.033758150602806857e-04,
-1.735652668867260261e-03,
--5.230193650641012590e-03,
--7.767367543419822877e-03,
-2.340998116414993915e-03,
--3.053004795219749104e-04,
-3.388495469698682341e-03,
-1.146385821048170363e-02,
-7.131879683583975380e-05,
--2.212631312431767583e-03,
-3.689170682264375500e-03,
--5.162494222167879690e-03,
--1.359342330833896925e-02,
-2.311418543104082346e-02,
--9.391274248628178611e-03,
--1.762565711978823085e-02,
-6.251880567288026040e-03,
-1.947507957083871588e-03,
-2.795461623463779688e-03,
-1.647698364831740080e-03,
-7.232262112665921450e-03,
-2.688910549040883849e-02,
--4.022772051393985401e-03,
--2.245790499728173018e-03,
-6.667657778598368515e-03,
--1.472272280079778327e-03,
--4.649207085640228235e-03,
--4.195724590681492935e-03,
--5.510665918700396668e-03,
--4.904550116043538202e-03,
--1.096505108580458909e-02,
--1.652644263813272257e-02,
--2.094466497437679278e-02,
-1.490428359829820693e-02,
-3.355966765666380802e-02,
--1.684766332618892193e-02,
--3.217324654106050452e-02,
-2.345243566378485238e-03,
-2.975666499696671963e-02,
--5.894004006404429501e-03,
--5.304902521311305116e-03,
-1.836623017443344066e-02,
--1.041659891052404419e-02,
--1.308841204736381839e-02,
-8.750846795737742684e-03,
--1.585985493147745772e-03,
--2.155492825659166556e-02,
--1.591869792900979519e-02,
-9.432580490829423417e-03,
--1.645074991974979778e-02,
--1.258636700222268666e-02,
--1.627706045983359079e-02,
--5.557938426500185977e-03,
-7.294263719813898206e-03,
--6.796524394303560257e-03,
--6.146338442340493549e-03,
--1.002743665594607592e-03,
--1.544136135489679838e-02,
--1.484100372981629265e-02,
-2.701981835998594830e-02,
-5.035066942218691458e-03,
--1.081432665814645658e-02,
--3.915473644156008959e-03,
-1.116530440631322528e-02,
--7.568762582377530461e-03,
--9.118141054932494063e-03,
-2.363307060295483097e-02,
--9.958223241847008120e-03,
--2.491156398318707804e-02,
-6.787785215419716960e-03,
-2.575042606913484702e-02,
--6.805236090440302914e-03,
-9.676721501455176969e-03,
-5.589777072891592702e-02,
-3.817702585365623019e-04,
--4.339076804462820014e-02,
-3.820459780399687480e-04,
--5.648350075352937330e-03,
--1.472614641534164461e-03,
-1.556033007800579002e-02,
-3.072406929350108770e-02,
--2.922106955156778159e-03,
--1.314932932145893539e-02,
-7.189174648374319348e-05,
--3.909531689714640379e-03,
-9.833761799382046559e-03,
--1.543392202001996408e-02,
--1.873548591975122621e-02,
-1.396073693467769815e-03,
-1.114964298030827160e-02,
--7.766276531037874173e-03,
--9.421445301268250444e-03,
-1.629893528297543526e-02,
-2.271096256736200233e-02,
-1.941514198260847410e-02,
--4.663562576752155748e-05,
--1.059749281557742506e-02,
--1.802444530185311894e-02,
--7.666371113919013061e-03,
--3.442160633858293208e-03,
-1.008582352660596440e-02,
-8.141988539136946201e-03,
-1.397803002037107875e-02,
-4.967673012288287146e-05,
--3.373269668518332626e-02,
--6.694578359019942755e-03,
--1.552248752559535171e-02,
-4.055497207446023822e-03,
--1.449119233657256665e-02,
-3.364010507357306952e-03,
--1.170082582393661098e-03,
--1.833085193356964737e-02,
--1.228844129946082830e-02,
-1.755480601277668168e-02,
-1.129883016692474415e-02,
--1.932738788309507115e-03,
--1.066097709408495540e-02,
--2.368945957277901384e-03,
--1.939881176222115616e-02,
-2.642210719641297986e-02,
-3.523698427015915791e-02,
-2.831683300610165870e-03,
-2.283063313225284358e-02,
-7.182254374492913138e-03,
-7.592860155273229057e-03,
--2.852862208383157711e-02,
--2.290775818109978129e-02,
-4.247987813141662454e-03,
--1.474833546672016382e-02,
-1.022195924306288327e-02,
--3.512530467705800991e-02,
-7.733812279184348881e-03,
--2.876970649231225165e-03,
--5.945696932030841708e-03,
-1.120745067601092160e-02,
-1.347201458411291172e-02,
--5.025910213589668545e-05,
-9.063865797361358306e-03,
--2.138226906536146860e-02,
-5.109271622495725404e-03,
-3.252034211764111976e-02,
-4.813811421627178788e-03,
-1.508184662670828367e-02,
--1.670899503224063803e-02,
-1.853182199411094327e-02,
--4.729678560397587890e-03,
-9.787266026251018047e-04,
--1.496538367355242441e-02,
--2.093779217102564698e-03,
--1.896673912415280895e-03,
--1.081579857855103841e-02,
-3.957933053607120763e-03,
--2.172633457230404089e-02,
--3.457329624798148593e-02,
-1.214827691728714810e-03,
-1.778758710133843124e-02,
-3.549763327755499354e-02,
--4.769720975309610714e-03,
-1.853583410411374643e-02,
-1.140166012919507921e-02,
--1.208045942476019229e-02,
-1.240584195766132256e-02,
-1.667359729763120274e-02,
-5.946851304133815748e-03,
--2.746045141248032277e-03,
--8.850714773871004581e-03,
--3.338876715861261064e-03,
--1.192847864003851975e-02,
-1.199488609127001906e-02,
-1.149435845436528290e-02,
--2.581064609694294631e-02,
--1.658009956590831419e-02,
--2.050778288958099438e-02,
--1.787968560383888100e-02,
-7.742021304147783320e-03,
--1.040473229659255681e-03,
-1.410851093241944840e-02,
-2.733687475847546098e-03,
--3.607173307682387357e-03,
--1.353686344955349306e-02,
--6.597327318741008466e-03,
--2.623887266963720457e-05,
-1.024312098161317408e-02,
-2.999159991741180351e-02,
--2.472464829916134338e-02,
-3.726926597300916823e-03,
--1.232063816860318184e-02,
--9.800773643655702125e-03,
-7.396876498023630293e-03,
--2.098957714333664695e-02,
-2.578735438110015682e-02,
-8.242677577072754502e-03,
-1.317472557420842384e-02,
-1.036536730607622372e-02,
-1.477463480114238337e-02,
-2.154917869338533049e-02,
--1.254921252839267219e-02,
-1.603017965353501634e-02,
--2.382105081342160771e-02,
-1.226091252174228360e-02,
--6.386895756440936003e-03,
--1.730546680919360300e-02,
-7.919914429658092211e-03,
-2.337965881451964378e-03,
-1.062291724374517764e-02,
--1.080543987627606908e-02,
--3.528470997116528365e-03,
-1.129181825090199744e-02,
-1.276588563341647317e-02,
--2.182105662723188115e-03,
--3.664668882265687097e-04,
-1.510079970175865968e-02,
--4.790010943543166269e-03,
--9.372872201493010619e-03,
-7.070297107566148390e-03,
--1.671674638055265037e-03,
-1.543382393429055741e-02,
-1.988150769830099134e-03,
--3.390072542242705649e-03,
-8.580265421187505126e-05,
-1.158945317147299682e-02,
-4.409580267383717347e-03,
-1.178605900495313143e-02,
--8.030706632416695703e-03,
--2.014953478937968681e-02,
--1.642542690387926996e-02,
--1.447836190345697105e-02,
-2.333743995986879011e-02,
-6.274123701587086327e-04,
--1.482019681716337765e-02,
--1.289150690427049925e-02,
-1.551784138428047299e-02,
--5.799001222476363529e-03,
--1.111309549378347583e-02,
-2.339902250096202035e-02,
-1.466969263856299222e-02,
-6.957293767482042139e-03,
--4.126671082340181013e-02,
--1.127030150964856113e-02,
--1.968681471189483861e-02,
-7.790977117838338191e-03,
--6.818551523610949690e-03,
-2.176870656403480028e-02,
-9.962659334996714841e-03,
--2.005878263007616555e-02,
--8.058609580621123314e-02,
--1.115404907613992691e-02,
-1.750798695284174616e-02,
-8.185246033826842557e-03,
--1.044064182788133691e-02,
-1.050373016914818436e-02,
--1.867606198356952562e-03,
-1.164274354814551736e-02,
-1.584685587440617244e-02,
--8.094696007901802659e-03,
-1.618924710637657111e-02,
-1.049668790155919816e-02,
--2.105879418668337236e-02,
-8.109498256817459366e-03,
-1.048479442833922734e-02,
--1.902036584215238647e-03,
-1.951753113535232773e-02,
-6.272327990154735397e-03,
--1.508215380599722301e-02,
--3.108658625569660130e-03,
--2.600858244113624009e-03,
-4.492764989117858584e-03,
-9.510822655283846655e-03,
-2.207101171952672442e-02,
--6.011746393050998102e-03,
-6.285632514118333174e-03,
--1.745754891890101065e-02,
-5.950855642186070500e-03,
-1.085624835395719925e-02,
--1.312971733859740267e-02,
-2.141740457154810498e-02,
-1.625992181361652841e-02,
-1.881551253609359264e-02,
-4.782891842478420412e-03,
-5.902485275873914708e-03,
--8.007584977895021786e-03,
--6.657903955783695199e-04,
--6.338556553123453688e-03,
-1.873688762134406716e-02,
--2.799125295132398519e-03,
--1.485318747327255516e-02,
--1.974958196515217562e-02,
-1.402924317517317832e-02,
-1.600786525523290038e-02,
-1.012879345798864926e-02,
--1.733242431073449621e-02,
-4.443230817560106689e-03,
-1.991378492966760019e-02,
--5.432185268728062337e-03,
-5.151906976243481387e-03,
-1.061795622299541685e-02,
--1.310527623281814227e-02,
-9.549642111960566365e-03,
--1.368443747633136739e-02,
--2.017398658499587258e-02,
--2.980093393489369300e-04,
--1.404808921506628394e-02,
--1.412533291440922802e-02,
--3.294883191119879658e-03,
-4.454243334475904530e-03,
--8.421283273492008448e-03,
-1.694765302818268494e-03,
--1.528159393928945134e-02,
--2.055281335487961838e-02,
-6.169875222258269960e-03,
--2.457632169534917921e-02,
--2.141837933158967713e-02,
-2.076982010694337216e-02,
--1.619636945833917777e-02,
-1.141766506480052998e-02,
-7.418549148133024920e-03,
-1.397676406195387175e-02,
-1.673910088138654689e-02,
-1.613335749425459653e-02,
-5.561829471844248474e-03,
--7.539799564983695919e-03,
--1.220674879732541740e-02,
-3.686573418235639110e-04,
--7.471651983360061075e-03,
--5.863940715789795269e-03,
-6.183539346966426993e-03,
-3.578870330238714901e-03,
-2.367767170071601729e-02,
-2.739032363751903176e-02,
--2.028330962639302015e-02,
-1.344616301357746090e-02,
-1.428120482014492119e-02,
--2.827200508909299883e-03,
--8.189644156482244960e-03,
--2.526918210787698453e-03,
-1.972217550501227309e-02,
-2.116161064186599175e-02,
-1.536526941927149857e-02,
--4.831511533120646726e-03,
-5.996225233684526583e-03,
-3.902642661705613050e-03,
--2.216451888671144843e-03,
--9.701133734779432416e-03,
-6.672447946039028119e-03,
--1.623004267457872629e-02,
-1.459162980318069389e-02,
-2.543707378208637238e-02,
-4.625371028669178659e-03,
--3.959999518701806838e-03,
--1.582958090584725072e-02,
--7.261072139954194427e-03,
-2.129947024368448322e-02,
--8.467121564899570893e-03,
-2.652767434483394027e-04,
--1.617802514228969951e-02,
-1.156479923520237207e-02,
--1.161675553303211962e-02,
-4.582954902434721763e-03,
-2.149930573068559170e-03,
-1.915007895149756223e-03,
-1.154905886651249637e-02,
--1.589805935509502779e-04,
-1.472745466744527186e-02,
-1.787011672277003665e-02,
--1.056974961029482143e-02,
--1.793417420412879437e-02,
-1.266242929116287962e-02,
--9.215665049850940704e-04,
--2.532801583583932370e-03,
--2.488375094253569841e-02,
--6.568782066460698496e-03,
-3.595062880776822431e-05,
--4.463918394321808591e-03,
--6.239975315111223608e-03,
-3.109409038734156552e-03,
-3.349567239056341211e-02,
-3.755253106355667392e-02,
--3.556570957880467003e-03,
--3.110190934967249545e-03,
-2.291007795720361243e-02,
--5.840362713206559767e-03,
-1.526814399985596503e-02,
-1.828139933058992028e-02,
--1.271463426528498565e-02,
-8.613508302005357109e-03,
-9.230293310247361313e-03,
--1.178385390085168270e-03,
-1.423493017064174707e-02,
--1.578779250849038363e-02,
--8.687859072233549904e-03,
-7.870240699048736566e-04,
--7.366210827603936542e-03,
-1.244723338168114354e-02,
--8.318259421503170631e-03,
--1.547621686841011909e-02,
-6.287893085391261030e-03,
-8.703948726179078655e-04,
-3.320142277516424656e-04,
-7.771216175751760429e-03,
--1.346900489879771998e-02,
-3.441873936681076607e-02,
-8.609205285029019467e-03,
--8.592266601044683955e-03,
--1.194944872113410349e-02,
-3.416566182859241685e-02,
--9.320688998559489152e-03,
-1.426526990471757052e-02,
--9.793337713927031257e-03,
--2.002020182553678612e-02,
-3.335717949084937572e-03,
-3.435405265918234363e-03,
--2.699632043950259824e-02,
--1.501216084870975534e-02,
--4.433116503059863697e-03,
-9.655935771297663797e-03,
-9.041206911206245422e-03,
--5.482473351366934540e-03,
-3.876987031253520313e-03,
--9.910093387588858604e-03,
-1.909288071328774138e-03,
-2.029564419353846461e-02,
-2.797715065535157780e-02,
--7.313537393929436976e-03,
--6.261164776515215462e-03,
-1.295709459591307643e-02,
--1.269880364998243793e-03,
-3.818508039694279519e-03,
--8.452152935569756317e-03,
-1.553340954123996233e-02,
-2.976664360612630775e-02,
--2.031826833263039762e-03,
-4.507955517619847974e-02,
--9.827088087331504324e-03,
--1.304875977803021756e-03,
-2.233754518965724856e-02,
-2.232093869242817233e-02,
-4.534311415045522327e-03,
-1.100584900705143740e-02,
--7.354353696064208981e-03,
-2.193159298331011017e-02,
-1.558988518372643808e-02,
--7.747381838271394372e-03,
-2.202932881773449550e-02,
-1.975725939264521072e-02,
--1.170127466320991559e-03,
-2.809884035377763212e-02,
-8.810946630546823144e-03,
-8.596165103517705922e-03,
--1.513510613585822371e-03,
--5.457887674856465657e-03,
-2.454112781379080843e-03,
-5.503082540235482503e-03,
-1.166542526334524155e-03,
--2.770771177019924064e-02,
--3.034130774904042309e-03,
--1.548818042938364739e-02,
-5.488219903782009740e-06,
-9.305982152000069618e-03,
-1.388136615569237654e-02,
--1.501197699690237728e-03,
--3.279524750541895801e-03,
--8.062720626185182482e-03,
-1.290465019410476019e-02,
--1.422986667603254318e-02,
--1.167788947532244406e-03,
-6.230182654689997435e-03,
--4.320573403965681514e-02,
--1.931539920624345680e-02,
-1.881197425827849795e-02,
--9.926602069754153490e-04,
-1.856734289322048426e-02,
--1.000952230242546648e-02,
--1.555628392670769210e-02,
-6.360858789412305106e-03,
-2.275860058143735001e-02,
--2.064108301419764844e-03,
-1.657381430268287797e-02,
-1.888478719629347463e-02,
--1.206091910135000979e-02,
-4.950877034571021622e-03,
--1.063337182276882266e-02,
-2.180212401435710576e-03,
-8.248750364873557589e-03,
-1.766643071314320107e-02,
--2.573822228587232391e-02,
-1.758153943810612102e-02,
--9.062338781586732478e-03,
-5.312807278824038630e-03,
--8.092578715877606771e-03,
-3.541209932882338676e-03,
--1.742221087915822933e-02,
-2.692934195511043072e-02,
-2.778443209826946328e-02,
--3.464140246069291780e-03,
-6.082908064490766611e-03,
-1.153738948050886349e-03,
--8.568021158862394043e-03,
-1.417300651155528612e-02,
-1.703941612504422665e-02,
-9.536510368343442332e-03,
--1.912544893275480674e-03,
-2.305097121279686689e-03,
--2.774106542346999049e-03,
--2.642214007209986231e-04,
-3.488037901115603832e-03,
--6.083788762043696070e-03,
-7.631018670508638024e-03,
-6.208615624927915633e-03,
--1.816763923707185452e-02,
--3.551857552884030069e-03,
-1.038769725710153580e-03,
--9.091123152757063508e-03,
-1.276774746511364321e-02,
--7.004351704381406090e-04,
--5.962423281744122072e-04,
--3.703397407662123333e-04,
-7.641337983659468765e-03,
-2.405025323969312012e-02,
-8.283065244904718946e-03,
--6.581717793596908624e-03,
-2.355621678289026091e-02,
-5.711935624276520684e-03,
--1.963596163550391724e-02,
--8.253962613525799713e-03,
--7.611531538350391549e-03,
--1.166138745102216455e-02,
-1.406958542356733302e-02,
--8.267057477496564171e-04,
--1.614632323980913454e-03,
-1.037310075189452549e-02,
--1.308949106896761805e-02,
--7.686795340850949634e-03,
--1.066315966309048201e-02,
-6.507713886094279587e-03,
--3.672932118934113706e-03,
-9.188450873989496143e-03,
--3.927565371122909363e-03,
-5.025976331671700228e-03,
-1.134529893752187456e-02,
-4.721692262683064092e-03,
--2.841148190200328688e-02,
--2.588801622623577645e-02,
-1.385693862685002421e-02,
-5.379268361139111221e-03,
-8.438213402405381550e-03,
-4.820020956685766578e-03,
-9.436098279547878304e-03,
--1.577339513460174203e-03,
-9.547394874971359697e-05,
-8.792690816335380077e-03,
-1.876135047568823142e-02,
--1.338109532371163299e-02,
--1.352524380490649492e-02,
--1.785166300833225181e-02,
--2.956867759348824447e-03,
-1.104491936857812141e-02,
--1.544391745701432193e-02,
-3.505717474035918799e-03,
-5.319875470013358046e-03,
-5.349700874648988420e-03,
--5.243656209131586365e-03,
-6.656189449131489017e-04,
-1.047522887820378015e-02,
-3.999249893240630627e-03,
--7.029033807339146560e-03,
--2.333188551710918632e-03,
--9.972017270047216914e-03,
--1.209601759910583496e-04,
-1.787121039742487622e-02,
--2.332056751474738052e-02,
--8.481634623603896000e-03,
--2.115292713278904644e-03,
--4.878579199521481877e-03,
-2.901614044094458396e-03,
-1.118667477858252893e-02,
--1.643372207763604820e-02,
--3.780025200103409762e-03,
-1.021935343160293962e-02,
--1.663857762468978719e-03,
-1.278010138776153326e-03,
-1.340722729510162095e-02,
-5.887076932413038040e-03,
-3.500899195205420429e-03,
--4.137214517686516220e-03,
--4.135751244029961179e-03,
--1.496108922874555061e-02,
--7.423543953336775303e-03,
-6.061168492306023663e-03,
-2.628208097303286021e-03,
--2.138131599640473793e-02,
-1.983559717773459782e-02,
--7.231167698046192180e-03,
-2.858428512990940189e-03,
--1.807744955876842280e-02,
--3.817988609080202964e-03,
-1.027301762253046008e-01,
--1.049182281713001517e-03,
--8.274902240373194218e-03,
-2.842782252810138570e-03,
-8.230978345090988677e-03,
--1.606798908323980911e-03,
-2.794570516562089237e-03,
--3.883002186194062320e-03,
-3.851585538359359080e-03,
-6.174380355514585798e-03,
--9.910328217665664505e-03,
--4.128728516661794856e-03,
--5.869911052286624908e-03,
-3.374529173015616834e-03,
-1.397572822752408614e-02,
--8.214152156142517924e-03,
--1.548986030975356767e-02,
--1.815933951293118367e-02,
--1.041212175914552063e-02,
-7.914100983180105339e-03,
-5.982818244956433600e-03,
--2.118119391961954465e-02,
-7.956661866046488285e-04,
-2.486425449023954744e-03,
-7.729447114979848092e-03,
--2.643775776959955692e-02,
--6.299247796414419662e-03,
-1.569755345117300818e-02,
-2.231216741347452626e-02,
-2.308206413872539858e-02,
--6.817292392952367880e-04,
-2.169419076744816353e-03,
--6.020447395712835714e-03,
--1.144130285538267310e-02,
--1.785701161716133495e-02,
--7.975140429334715703e-03,
--1.460391271893968224e-02,
--1.315491588611621479e-02,
--1.715717080514878102e-05,
--2.053622798412107001e-02,
-8.538523814058862973e-03,
-1.342314690118655632e-02,
-1.024173995892851963e-02,
--4.919437650823965495e-03,
-4.940598583198152136e-03,
--3.211722841951995810e-02,
--1.490932577289640938e-02,
--4.623951171379303386e-03,
-2.821497619152069092e-03,
--2.649463398847729054e-03,
--1.405036709038540686e-02,
-1.470845625153742800e-02,
-8.838723394364932789e-03,
--1.557127897103782700e-03,
--9.210367813648191043e-03,
-5.304105958202854016e-04,
--2.531161342631094158e-03,
--1.434937049634754701e-03,
-3.703486156882718205e-02,
-1.880257999291643639e-02,
--1.643821003381162925e-03,
--7.333502569235861128e-03,
-9.127936000004411091e-03,
--7.647617696784436356e-03,
-6.739112496143206607e-03,
--7.894508977187797644e-04,
--1.318564581597456674e-02,
--1.248691959772258928e-02,
-6.605757673969492490e-03,
-4.590558167546987360e-03,
--6.306250725174322724e-03,
-7.443017896730452602e-03,
-2.741795399924740022e-03,
--8.752206584904342196e-03,
--7.495709529030137111e-03,
-1.216502844472415815e-02,
--4.111754623590968376e-03,
-4.453295689017977457e-04,
--1.709070159122347971e-02,
-1.102580450242385268e-02,
-2.065254080516751994e-02,
--8.239158484502694807e-03,
-9.233326080720870993e-03,
-1.985971657559275697e-02,
--2.395824002451263542e-03,
-4.743019084708066310e-03,
-6.525630661781178614e-03,
--2.065923289046622813e-02,
--1.649279161938466132e-03,
--1.783575213630683720e-02,
-4.754915452213026757e-03,
-9.640499942906898684e-03,
-1.249273929279297660e-02,
-2.034188253455795209e-02,
--5.028525445959530424e-03,
--3.322070196736603889e-03,
--1.645722325192764540e-02,
-6.581127032404765704e-04,
-7.320888177491724665e-03,
-7.856722714495844445e-03,
-4.388645774452015584e-03,
--4.550904226198327143e-03,
--4.750136411166750242e-03,
-8.064571651630103588e-03,
-1.402271989500150119e-02,
--1.569007054786197841e-02,
--1.265118458541110089e-02,
--7.055002311244606798e-03,
-6.481654668459668582e-03,
-6.615672711632214487e-03,
-2.189354971051216125e-02,
-5.599562628776766533e-03,
-3.221723822643980661e-02,
--1.335754067185917007e-02,
-2.201184382771543754e-02,
--7.557125013408949772e-03,
--5.908043868839740580e-03,
--6.321127594856079494e-03,
--1.114656733989249915e-02,
--3.561108367284759792e-03,
-2.480515344068407943e-02,
--1.113502854714170168e-02,
-1.354382369900122320e-02,
-1.089393672300502727e-02,
-3.519720117037650024e-03,
--8.801410032901913361e-04,
-3.806073726445902226e-03,
-2.823798181489110132e-02,
-3.081401831004768749e-02,
--2.284305938519537622e-03,
-1.698123157257214145e-03,
--9.708647181105334659e-03,
-5.234614625805988908e-03,
--4.274043971963692630e-03,
-1.920406317804008653e-02,
-2.079129822668619385e-02,
--5.977447413670233800e-03,
--1.238779033010359819e-02,
--6.467461201827973474e-03,
--1.018362818285822868e-04,
-3.551157281617633825e-03,
--5.910143864821293218e-03,
--5.956706445431336761e-05,
--1.121888507623225451e-02,
-8.594607788836583495e-03,
-2.071786423912271718e-02,
-4.774379002628848130e-03,
-1.278833637479692784e-03,
--1.634748526412295108e-02,
--1.983580389060080051e-02,
--6.526824960019439459e-03,
-1.258983963634818836e-03,
--1.009841045597568120e-02,
-6.983118969947100033e-03,
--1.094209363654954394e-02,
-5.126653431216254657e-03,
--2.888724498916417360e-03,
--1.769471401348710017e-03,
-2.839042840641923388e-02,
--3.108765103388577526e-03,
-6.882070490974001586e-03,
-7.991266949102283218e-03,
--8.260777910800242915e-03,
-1.065625900264421944e-02,
-3.254883131012320518e-02,
-1.444831169646931856e-03,
--1.047802006360143423e-02,
--2.441555096447700721e-03,
--2.320154837070731359e-03,
--1.131398851866833796e-02,
-6.709687170223332595e-03,
-8.056852873414755214e-03,
--1.638028165325522510e-03,
--6.307428656145930464e-03,
-7.085928921151207413e-03,
--6.061694107484072273e-03,
-1.377297842191183095e-03,
-2.058515690441709012e-02,
--7.517418754287063902e-03,
--6.791737818275578320e-03,
--9.608150122221558592e-03,
--1.765466118667973275e-02,
-1.171345178736373817e-02,
--7.566322710772510618e-04,
--2.569345899391919236e-02,
-8.358489390229806998e-03,
-4.053028940688818693e-03,
--7.538855471648275679e-03,
--5.413519719149917017e-03,
--8.694138034479692231e-03,
-2.115376177243888378e-03,
--1.046781857730820844e-02,
-1.387932138750329683e-02,
-1.354039473408192792e-02,
--1.206661026226356695e-02,
-2.293357864255085588e-02,
--3.610883929650299078e-03,
--1.243656184815336053e-02,
-4.586090275552123785e-03,
-5.200736594269983123e-03,
-7.774297015566844146e-03,
--1.043932152679189970e-02,
-1.743367178132757683e-02,
-8.978349773678928267e-03,
-2.936053164303302834e-02,
--1.515399164054542693e-03,
-1.559907675255089998e-02,
-2.093702155398204784e-02,
--2.587611760827712541e-03,
-9.490785218076780438e-03,
-1.903583339299075305e-02,
--9.700666961725801576e-03,
-1.582571490143891467e-03,
--1.267328932590316894e-03,
-2.362947221263311717e-03,
-1.946196645149029722e-02,
-9.011766636103858003e-03,
-9.982516494346782565e-03,
--1.225340649543795769e-02,
--9.711418789811432709e-03,
--1.911661276244558307e-03,
--7.381553962477483029e-03,
--8.592037385824369700e-04,
--2.242713966406881948e-02,
--1.917882619891315629e-02,
--1.247785883024334942e-02,
-1.914953821687959048e-03,
--8.775244912249036480e-03,
-3.990936730406247253e-03,
--3.856583044398576195e-03,
--2.273750600579660386e-03,
--1.436085702443961123e-02,
--4.667493810120504523e-03,
--4.294897755698912079e-04,
-2.130453842692077090e-02,
--8.692576014436782100e-03,
--1.817197035998106142e-02,
--6.492670497391372746e-03,
--7.498866040259599339e-03,
--5.412057670127978998e-04,
--1.215554334339685695e-03,
--3.073696017963811755e-03,
-1.334991632029414264e-03,
--3.550659447500948036e-03,
--9.801604400854558100e-03,
-2.646082234568893771e-02,
-1.637801051838323405e-02,
--1.874335035681724479e-02,
--8.882203121902420423e-03,
-1.042715916992165148e-02,
--8.008928503841161945e-04,
-3.445437260961625552e-03,
--2.066332043614238674e-03,
--1.658553721281350729e-02,
--4.701998946984531358e-03,
--2.336643333546817129e-03,
-5.569492704307776702e-03,
--6.626862176926806396e-03,
-6.663641403429209839e-03,
-2.671076881233602675e-03,
-4.870861949166282664e-03,
-2.109902994707226614e-02,
-2.374765824060887098e-02,
--1.163802237570052932e-02,
-2.233360558748245100e-02,
-3.538694733288139192e-03,
--2.363352337852120345e-04,
-1.954562460596207668e-02,
--5.183506470348220492e-03,
-5.724921787623316397e-03,
-8.219848998123779893e-03,
-6.348441937007009549e-04,
--5.559814680600539197e-03,
-1.281951499549904838e-02,
--1.220506809186190297e-02,
-1.891820076853036742e-02,
--8.371846610680222164e-03,
--1.943286275491118431e-03,
-3.899429836019408045e-03,
-1.576097751240013167e-03,
-7.976570978644303284e-03,
-4.187518160324543888e-03,
-6.003536604112014174e-03,
--9.083423356059938669e-03,
-1.936828077305108309e-02,
-2.860336349112912839e-03,
--1.972673910204321077e-02,
-1.472707611828809673e-02,
--3.324280976812588159e-04,
--2.488546670065261424e-03,
--5.497445724904537548e-03,
-1.623212448321282794e-02,
-2.072620307444594800e-02,
--1.351172099821269443e-02,
-1.657902354490943184e-02,
-8.916535173193551253e-03,
--5.132115678861737251e-03,
--1.732679416454629973e-02,
--1.084842742420732888e-03,
--4.556570958811789643e-04,
--1.681491440685931732e-02,
--8.767609047936275601e-03,
--4.238943082600599328e-03,
--1.592217848738073510e-02,
-7.929830462671816349e-03,
-8.896900151739828289e-03,
--4.933272162452340473e-03,
-1.622654025995871191e-02,
--6.320005346788093101e-03,
--4.353146724861289493e-03,
-6.533471448346972292e-03,
--4.855240543838590730e-03,
-2.668016718234867031e-03,
-1.574495651530014723e-02,
--6.509712709521409471e-03,
--5.505113236722536385e-03,
-2.978651813464239240e-03,
-1.452154327853349969e-02,
-1.062976232788059860e-02,
-1.576894427998922626e-02,
-2.189539915416389634e-02,
--3.995549643877893860e-03,
-3.494401695206761577e-04,
--6.171266562887467444e-03,
-1.223825230845250120e-02,
--2.051482626702636619e-02,
--6.103948038071393793e-03,
-1.262372150813462213e-02,
-2.649483201093971729e-02,
-2.597295416053384612e-02,
-5.615037761162966663e-03,
-1.238398958230391231e-03,
--5.837172879546414714e-03,
--3.033868223428726370e-03,
-2.823848689149599696e-03,
--4.485397596727125520e-03,
-8.510926156304776669e-03,
-3.129494094755500694e-02,
-3.058735091326525261e-03,
-7.470455323345959360e-03,
-3.947153632179833775e-03,
-1.762350695207715034e-02,
-3.305345369153655892e-03,
--9.971052647051692822e-03,
-2.668607260129647260e-02,
-2.279791150358505616e-02,
-1.376637805224163388e-02,
--3.190379391890019265e-03,
--1.087475954118417507e-02,
--5.232337198685854326e-03,
--1.684350024443119834e-02,
--1.514592983294278314e-02,
--8.012250671163201679e-03,
-5.143893556669354786e-03,
-8.547383337281643997e-03,
-1.025208962091710488e-02,
-2.076280133333057235e-02,
--3.050260457675903897e-02,
--3.929960570530965924e-03,
-7.651075691683217883e-03,
--6.275656138313934424e-04,
-8.716393106442409805e-03,
-6.821079288056352934e-03,
--5.091704278311226343e-03,
-1.078921162988990580e-02,
-3.676004935987293998e-02,
--8.612308790907264103e-03,
--7.157141552306711327e-03,
--6.438640132546425039e-03,
--3.417930004070513037e-03,
--1.059016523067839513e-02,
-3.383215147187001916e-03,
--1.310363429365679665e-03,
--1.758118880534311618e-02,
-2.436707203742116744e-04,
--1.667125566746108147e-03,
--1.304479663156143943e-02,
--1.887810298940166900e-02,
-4.702559587894938713e-03,
--9.527037036605179310e-04,
-1.051573613367509132e-02,
--1.535475985147058929e-02,
-8.621049416251479885e-03,
--7.776543556246906519e-03,
-1.011012516682967613e-02,
-1.165708121552597697e-02,
--4.989801449846708800e-03,
--8.666561171412467263e-03,
-1.671402715146541595e-04,
-    ],
-  }
 
   async function get_text_embedding(text) {
     const response = await fetch('https://api.openai.com/v1/embeddings', {
@@ -4669,46 +57,14 @@ function OverleafCursor() {
   async function handle_autocomplete(focused_text, around_text, all_text) {
     if (all_text == "") return "";
 
-    // check if the user is asking for an autocomplete
-    // const validation = getCheckAutocompleteNeededPrompt(focused_text);
-    // let validation_response = await hitOAI([{role: 'user', content: validation}]);
-    // validation_response = validation_response.replace('```json', '').replace('```', '');
-
     // get the embedding of the focused text
     const focused_embedding = await get_text_embedding(focused_text);
-    
     // get the cosine similarity between the focused embedding and the average instructional creational embedding
-    const cosine_similarity = get_cosine_similarity(focused_embedding, average_vectors["instructional_creational"]);
-    console.log(cosine_similarity);
-    
-    
-    // try {
-    //   // Try to parse the JSON response
-    //   validation_response = JSON.parse(validation_response);
-    // } catch (error) {
-    //   console.error('Error parsing JSON response:', error);
-    //   // If parsing fails, try to extract the answer using regex
-    //   const answerMatch = validation_response.match(/"answer"\s*:\s*"([^"]+)"/);
-    //   const reasoningMatch = validation_response.match(/"reasoning"\s*:\s*"([^"]+)"/);
-      
-    //   if (answerMatch && reasoningMatch) {
-    //     validation_response = {
-    //       answer: answerMatch[1],
-    //       reasoning: reasoningMatch[1]
-    //     };
-    //   } else {
-    //     // If we can't extract the answer, default to "no"
-    //     validation_response = {
-    //       answer: "no",
-    //       reasoning: "Failed to parse response"
-    //     };
-    //   }
-    // }
-    
-    // console.log('Validation Response:', validation_response);
-    
-    // if (validation_response["answer"] === "yes") {
-    if (cosine_similarity > 0.5) {
+    const similarities = similarity_all(focused_embedding);
+    console.log(similarities);
+
+    let largest_similarity = Math.max(similarities.map(similarity => similarity.similarity));
+    if (largest_similarity > 0.4) {
       const prompt = getAutoCompletePrompt(all_text, around_text);
       let response = await hitOAI([{role: 'user', content: prompt}]);
       response = response.replace('```latex', '').replace('```', '').trim().replace('`', '').replace('"', '');
@@ -4762,89 +118,6 @@ function OverleafCursor() {
     }
   });
 
-}
-// PROMPTS ----------------------------------------------------------------------
-function getEditPrompt(code_snippet, user_request) {
-  return `
-  You are a LaTeX expert. You are given a LaTeX code snippet and a user request.
-  You need to edit the LaTeX code snippet to fix the errors and improve the code.
-  Keep the code clean, short, and concise.
-
-  Here is the LaTeX code snippet:
-  ${code_snippet}
-
-  Here is the user request:
-  ${user_request}
-
-  RETURN ONLY THE LATEX CODE AS A STRING. INCLUDE \n BETWEEN LINES OF CODE. 
-  `
-  }
-
-
-function get_cosine_similarity(vector1, vector2) {
-  // calculate the dot product of the two vectors
-  let dot_product = 0;
-  for (let i = 0; i < vector1.length; i++) {
-    dot_product += vector1[i] * vector2[i];
-  }
-  
-  // calculate the magnitude of the two vectors
-  let magnitude1 = 0;
-  let magnitude2 = 0;
-  for (let i = 0; i < vector1.length; i++) {
-    magnitude1 += vector1[i] * vector1[i];
-    magnitude2 += vector2[i] * vector2[i];
-  }
-  magnitude1 = Math.sqrt(magnitude1);
-  magnitude2 = Math.sqrt(magnitude2);
-  
-  // calculate the cosine similarity
-  const cosine_similarity = dot_product / (magnitude1 * magnitude2);
-  return cosine_similarity;
-}
-
-  // function getCheckAutocompleteNeededPrompt(focused_text) {
-//   return `
-//   You are a LaTeX expert. Based on the LaTeX codebase and highlighted area using dashes, determine if the an autocomplete is needed.
-//   Return "yes" if the an autocomplete is needed, otherwise return "no". Do not repeat code that is nearby the highlighted area.
-//   Be very greedy with how many times you return "yes". 
-
-//   Here is the highlighted area:
-//   ${focused_text}
-
-//   Only return "yes" if you are very confident that a meaningful and relevant autocomplete is needed. Say NO more often than yes.
-//   RETURN JSON IN THE FOLLOWING FORMAT:
-
-//   {{
-//      "reasoning": "reasoning for your answer",
-//      "answer": "yes" or "no"
-//   }}
-// `
-
-
-  // hardcoded groups of vectors for types of autocomplete (instructional, syntactical, etc)
-  // use cosine similarity to determine which group the user request most closely aligns with
-  // return the group that has the highest cosine similarity
-  
-
-  
-  
-// }
-
-
-function getAutoCompletePrompt(all_text, plusminus_text) {
-  return `
-  The code provided below is a snippet of a larger LaTeX codebase. Assume not all the code is provided.
-  Complete the latex code in only the lines provided. Do not create new lines of code.
-
-  Here is the LaTeX codebase:
-  ${all_text}
-
-  Complete the following latex code:
-  ${plusminus_text}
-
-  RETURN ONLY THE LATEX CODE PROVIDED WITH THE COMPLETION CODE.
-  `
 }
 
 // Initialize the extension
